% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArbitraryCover.R
\docType{class}
\name{ArbitraryCover}
\alias{ArbitraryCover}
\title{Arbitrary Cover}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ArbitraryCover
}
\description{
This class enables the user to use any fixed arbitrary cover. 
To facilitate this, rather than providing a parameterized family of covers as the others
do (see e.g. \code{\link{FixedIntervalCover}}), the user provides the indexed set of 
open sets themselves.
}
\seealso{
Other cover: \code{\link{BallCover}},
  \code{\link{FixedIntervalCover}}
}
\author{
Matt Piekenbrock
}
\concept{cover}
\keyword{datasets}
