% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{as_grapher}
\alias{as_grapher}
\title{Exports the 1-skeleton as a grapher object.}
\arguments{
\item{construct_widget}{whether to construct the htmlwidget or just the grapher configuration.}
}
\description{
Uses the grapher library.
}
