% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{as_pixiplex}
\alias{as_pixiplex}
\title{Exports the complex as a pixiplex object.}
\description{
Uses the \pkg{pixiplex} library.
}
\seealso{
Other mapper visualization functions: \code{\link{MapperRef.plot}},
  \code{\link{as_igraph}}
}
\concept{mapper visualization functions}
