% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{assemble_mapper_object}
\alias{assemble_mapper_object}
\title{Mapper Object Assembler}
\usage{
assemble_mapper_object(binclust_data, dists, binning = TRUE)
}
\arguments{
\item{binclust_data}{A list of lists, each containing named vectors whose names are those of data points and whose values are cluster IDs (integers).}

\item{dists}{A distance matrix for the data that has been binned and clustered. Can be a \code{dist} object or a \code{matrix}.}

\item{binning}{Whether the output data frame should sort vertices into "patches" or not. Should be true if using clustering, leave \code{FALSE} otherwise (for, say Ball Mapper).}
}
\value{
A \code{list} of two data frames, one with node data and one with edge data. The node data includes:
\itemize{
\item \code{id}: vertex ID
\item \code{cluster_size}: number of data points in cluster
\item \code{mean_dist_to_medoid}: mean distance to medoid of cluster
\item \code{data}: names of data points in cluster
\item \code{patch}: level set ID (if \code{binning} was \code{TRUE})
}

The edge data includes:
\itemize{
\item \code{source}: vertex ID of edge source
\item \code{target}: vertex ID of edge target
\item \code{weight}: Jaccard index of edge; intersection divided by union
\item \code{overlap_data}: names of data points in overlap
\item \code{overlap_size}: number of data points overlap
}
}
\description{
Mapper Object Assembler
}
