% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapper-package.r
\name{bin_color}
\alias{bin_color}
\title{bin_color}
\usage{
bin_color(x, col_pal = rev(rainbow(100L, start = 0, end = 4/6)),
  output_format = c("hex9", "hex7"))
}
\arguments{
\item{x}{A numeric vector whose magnitudes should be binned onto the color palette.}

\item{col_pal}{Color palette to bin numeric values into. See details.}

\item{output_format}{Whether the output should be RGBa (hex9) or RGB (hex7).}
}
\description{
The \code{col_pal} must be a string vector of hexadecimal RGB or RGBa color codes.
}
\details{
Given a numeric vector \code{x}, bins the values from low to high on a given color gradient. 
Defaults to the reversed rainbow gradient, where blue == low, red == high.
}
