% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{brushScatter}
\alias{brushScatter}
\title{brushScatter - shiny server module}
\usage{
brushScatter(input, output, session, mapper_vis, X, xvar, yvar, reset)
}
\description{
brushScatter - shiny server module
}
