% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilting_hut.R
\name{check_in_interval}
\alias{check_in_interval}
\title{Get a tester function for an interval.}
\usage{
check_in_interval(endpoints)
}
\arguments{
\item{endpoints}{A vector of interval endpoints, namely a left and a right. Must be in order.}
}
\value{
A function that eats a data point and outputs TRUE if the datapoint is in the interval and FALSE if not.
}
\description{
Get a tester function for an interval.
}
