% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{compute_edges}
\alias{compute_edges}
\title{Computes the edges of the mapper.}
\arguments{
\item{which_level_pairs}{something}

\item{min_weight}{something2}
}
\description{
Computes the edges composing the topological graph (1-skeleton).
}
