% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{compute_vertices}
\alias{compute_vertices}
\title{Computes the vertices of mapper.}
\description{
Executes the clustering algorithm for the level sets indexed by the 'which_levels' parameter. If not given, 
runs the clustering algorithm and computes the subsequent vertices for all the available level sets. Additional 
parameters passed via the '...' are passed to the clustering algorithm.
}
