% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_mapper.R
\name{construct_1Dmappergraph}
\alias{construct_1Dmappergraph}
\title{Run 1D mapper algorithm!}
\usage{
construct_1Dmappergraph(binclust_data, dists)
}
\arguments{
\item{binclust_data}{A list containing named vectors whose names are datapoint names and whose values are cluster labels.}

\item{dists}{A distance matrix containing pairwise distances between named datapoints.}
}
\value{
A list containing:
\itemize{
\item nodes - A dataframe containing vertex/cluster information:
\itemize{
\item id - The vertex label.
\item size - How many datapoints are contained in that vertex.
\item tightness - A measure of dispersion of the data inside the vertex.
\item data - A string containing names of the datapoints within the vertex.
\item bin - Which bin the vertex belongs to.
}
\item edges - A dataframe containing vertex/cluster overlap information:
\itemize{
\item source - One endpoint of an edge.
\item target - The other endpoint of an edge.
\item weight - A measure of the significance of the edge.
}
}
}
\description{
Run 1D mapper algorithm!
}
