% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{construct_annotation}
\alias{construct_annotation}
\title{Compute simplex annotations}
\description{
Constructs a data frame of simplex-level annotations of the
  mapper skeleton, to be used for plotting.
}
\details{
This method calculates several simplex-level variables for use with
  the \code{\link[base:plot]{$plot()}} method:
\describe{
  \item{\code{dimension}}{the dimension of the simplex}
  \item{\code{size}}{the number of observations represented in the simplex}
  \item{\code{color}}{the stain obtained using the chosen
                      \code{\link[use_palette]{palette}}}
}
}
