% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{construct_coordinates}
\alias{construct_coordinates}
\title{Fix vertex coordinates}
\arguments{
\item{layout}{passed to \code{$use_layout()}.}

\item{...}{passed to \code{$use_layout()}.}
}
\description{
Computes (if necessary) and saves a vertex-by-2 matrix of
  coordinates that will be passed by default to \code{$plot()}.
}
\details{
(Add details.)
}
