% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{construct_cover}
\alias{construct_cover}
\title{Constructs the cover on the codomain of the map.}
\arguments{
\item{index}{indices of the \code{\link[Mapper:CoverRef]{covers}} \code{index_set}, or \code{NULL}.}

\item{...}{additional parameters to pass to the \code{construct} method of the cover.}
}
\description{
Executes the clustering algorithm on subsets of \code{X},
}
\details{
Constructs the cover over the codomain of the filter function.
}
