% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{construct_k_skeleton}
\alias{construct_k_skeleton}
\title{Constructs the k-skeleton}
\arguments{
\item{k}{the maximal dimension to consider.}
}
\description{
Computes the k-skeleton of the mapper by computing the nerve of the 
the pullback cover induced by the \code{cover} member. \cr
\cr
This function computes the k-skeleton inductively, e.g. by first computing the vertices, 
then the edges, etc. A check is performed to ensure the pullback has been decomposed, and 
if not, then \code{\link{construct_pullback}} is called.
}
\details{
For the details on how this is computed, see Singh et. al, section 3.2.
}
