% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{construct_nerve}
\alias{construct_nerve}
\title{Compute the nerve of the cover.}
\arguments{
\item{k}{The order of the simplices to construct. See details.}

\item{indices}{(n x k) matrix of indices of the covers index set to update.}

\item{min_weight}{minimum intersection size to consider as a simplex. Defaults to 1.}
}
\description{
Computes (or updates) the k-simplices composing the Mapper, where k >= 1.
}
\details{
Compared to \code{construct_k_skeleton}, this method \emph{only} intersections 
between (k-1) simplices in the complex.
}
