% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{construct_pullback}
\alias{construct_pullback}
\title{Computes the vertices of the mapper at a given index.}
\description{
Executes the clustering algorithm for sets indexed by \code{pullback_ids}.
}
\details{
If ids are not supplied, the pullback is computed for all the sets in the cover.
parameters passed via the '...' are passed to the clustering algorithm. \cr
\cr
Note that this method removes the vertices associated with \code{pullback_ids} from the simplicial 
complex, which subsequently also any of their cofaces. New vertices computed with \code{clustering_algorithm}
have their ids generated by \code{generate_ids}.
}
