% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_factory.R
\name{convert_to_clusters}
\alias{convert_to_clusters}
\title{The easiest clustering method}
\usage{
convert_to_clusters(bins)
}
\arguments{
\item{bins}{A list of bins, each containing names of data from some data frame.}
}
\value{
A named vector whose names are data point names and whose values are cluster labels
}
\description{
The easiest clustering method
}
