% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{cover}
\alias{cover}
\title{Mapper cover}
\description{
Every \code{\link{MapperRef}} object requires a \code{\link{CoverRef}} object as 
is \code{cover} member field. In the context of Mapper, a cover is used to discretize the filter 
space into a partition, which is then used via a \emph{pullback} operation to construct the vertices. \cr 
\cr 
The \code{\link{MapperRef}} class makes no restrictions on the cover that is used; only that it fulfills the 
requirements of being a valid \code{\link{CoverRef}} instance.
}
\seealso{
\code{\link{CoverRef}}
}
