% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_ref.R
\docType{class}
\name{CoverRef}
\alias{CoverRef}
\title{Cover reference class - Reference Class (R6) implementation of a Cover}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CoverRef
}
\description{
Cover reference class - Reference Class (R6) implementation of a Cover
}
\section{Fields}{

\describe{
\item{\code{filter_values}}{(n x d) matrix of filter values}

\item{\code{type}}{character string of the type of cover}
}}

\author{
: Matt Piekenbrock
}
\keyword{datasets}
