% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_methods.R
\name{cutoff_first_bin}
\alias{cutoff_first_bin}
\title{cutoff_first_bin}
\usage{
cutoff_first_bin(hcl, diam, num_bins)
}
\arguments{
\item{hcl}{hierarchical clustering in the form an 'hclust' object.}

\item{diam}{largest pairwise distance between the objects used to create the hierarchy.}

\item{num_bins}{Controls how many bins there are in the histogram used to determine cutoff.}
}
\description{
Uses the heuristic described in Section 3.1 of the original Mapper paper to cut a given hierarchical tree to produce a partitioning of the data.
The cut value is chosen as the lowest break point corresponding to the first empty bin of a histogram of the linkage distances.
}
