% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_methods.R
\name{cutoff_first_threshold}
\alias{cutoff_first_threshold}
\title{cutoff_first_threshold}
\usage{
cutoff_first_threshold(hcl, threshold = 0, ...)
}
\arguments{
\item{hcl}{hierarchical clustering in the form an 'hclust' object.}

\item{threshold}{density threshold.}

\item{...}{Additional parameters passed to \code{\link[stats]{density}}.}
}
\description{
Uses a heuristic similar to the one described in Section 3.1 of the original Mapper paper to cut a given hierarchical tree to produce a partitioning of the data.
The cut value is chosen as the lowest (interpolated) height value where the density of linkage distances is lower than or equal to 
a given threshold value.
}
