% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballmapper.R
\name{cyballmapper}
\alias{cyballmapper}
\title{Runs ballmapper and passes data to Cytoscape for visualization.}
\usage{
cyballmapper(data, dists, eps)
}
\arguments{
\item{data}{Your input data. Ideally a dataframe.}

\item{dists}{A distance matrix for your data. Can be a \code{dist} object or 2D matrix.}

\item{eps}{A positive real number for your desired ball radius.}
}
\description{
Runs ballmapper and passes data to Cytoscape for visualization.
}
