% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{distance}
\alias{distance}
\title{Computes pointwise distances}
\arguments{
\item{x}{numeric matrix of points.}

\item{y}{optional numeric matrix of points to compare to \code{x}. See details.}

\item{...}{Other parameters passed to \code{\link[proxy]{dist}}.}
}
\description{
This function computes distances using the given distance measure. See details.
}
\details{
This function mimicks the \code{\link[proxy:dist]{proxy distance function}}, but 
uses the \pkg{parallelDist} package when possible. Specifically, if \pkg{parallelDist} is available, 
only \code{x} was supplied (implying all self-pairwise distances are requested), and the 
\code{\link{measure}} is supported, then the distances are computed in parallel. Otherwise, the 
\pkg{proxy} package's \code{\link[proxy]{dist}} function is called. \cr
\cr
The convention followed by this function that is each row represents a point, 
and each column a dimension.
}
