% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{distance_subset}
\alias{distance_subset}
\title{Extract subset of distances}
\arguments{
\item{idx}{Indices in the range [1, n].}
}
\value{
a \code{\link[stats]{dist}} object.
}
\description{
Returns the distances between all pairs of points
}
