% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{exportMapper}
\alias{exportMapper}
\title{Exports the essential mapper graph information}
\arguments{
\item{graph_type}{export preference on the structure the graph output.}
}
\value{
list with the following members: 
\itemize{
\item vertices list of the indices of the original data the current vertex intersects with.
\item graph some adjacency representation of the mapper graph.
\item level_sets map connecting the which vertices belong to the preimage of the sets in the cover. 
}
}
\description{
Exports the essential mapper graph information
}
