% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedIntervalCover.R
\docType{class}
\name{FixedIntervalCover}
\alias{FixedIntervalCover}
\title{Fixed Interval Cover}
\description{
\code{\link{R6}} class representing a fixed interval cover.
}
\details{
The fixed interval cover is multidimensional, two-parameter family of covers. Given the number of
intervals and the overlap percentage between these intervals, this class constructs cover of the
given filter space, where the open sets represent a (generalized) collection of interval level sets, 
distributed uniformly in a grid-like fashion. Unlike the \code{\link{RestrainedIntervalCover}}, the open sets
are not constrained by range of the filter values. \cr 
\cr 
The mapper corresponding to this cover may be thought of as a relaxed Reeb graph.
}
\author{
Matt Piekenbrock
}
\concept{cover}
\section{Super class}{
\code{\link[Mapper:CoverRef]{Mapper::CoverRef}} -> \code{FixedIntervalCover}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{number_intervals}}{vector of number of bins to cover the Z with (per dimension)}

\item{\code{percent_overlap}}{vector of overlap percentages}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FixedIntervalCover-new}{\code{FixedIntervalCover$new()}}
\item \href{#method-FixedIntervalCover-validate}{\code{FixedIntervalCover$validate()}}
\item \href{#method-FixedIntervalCover-format}{\code{FixedIntervalCover$format()}}
\item \href{#method-FixedIntervalCover-interval_bounds}{\code{FixedIntervalCover$interval_bounds()}}
\item \href{#method-FixedIntervalCover-construct_index_set}{\code{FixedIntervalCover$construct_index_set()}}
\item \href{#method-FixedIntervalCover-construct_cover}{\code{FixedIntervalCover$construct_cover()}}
\item \href{#method-FixedIntervalCover-neighborhood}{\code{FixedIntervalCover$neighborhood()}}
\item \href{#method-FixedIntervalCover-overlap_to_interval_len}{\code{FixedIntervalCover$overlap_to_interval_len()}}
\item \href{#method-FixedIntervalCover-interval_len_to_percent_overlap}{\code{FixedIntervalCover$interval_len_to_percent_overlap()}}
\item \href{#method-FixedIntervalCover-clone}{\code{FixedIntervalCover$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedIntervalCover-new"></a>}}
\if{latex}{\out{\hypertarget{method-FixedIntervalCover-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedIntervalCover$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedIntervalCover-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FixedIntervalCover-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedIntervalCover$validate(filter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedIntervalCover-format"></a>}}
\if{latex}{\out{\hypertarget{method-FixedIntervalCover-format}{}}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedIntervalCover$format(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedIntervalCover-interval_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-FixedIntervalCover-interval_bounds}{}}}
\subsection{Method \code{interval_bounds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedIntervalCover$interval_bounds(filter, index = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedIntervalCover-construct_index_set"></a>}}
\if{latex}{\out{\hypertarget{method-FixedIntervalCover-construct_index_set}{}}}
\subsection{Method \code{construct_index_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedIntervalCover$construct_index_set(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedIntervalCover-construct_cover"></a>}}
\if{latex}{\out{\hypertarget{method-FixedIntervalCover-construct_cover}{}}}
\subsection{Method \code{construct_cover()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedIntervalCover$construct_cover(filter, index = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedIntervalCover-neighborhood"></a>}}
\if{latex}{\out{\hypertarget{method-FixedIntervalCover-neighborhood}{}}}
\subsection{Method \code{neighborhood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedIntervalCover$neighborhood(filter, k)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedIntervalCover-overlap_to_interval_len"></a>}}
\if{latex}{\out{\hypertarget{method-FixedIntervalCover-overlap_to_interval_len}{}}}
\subsection{Method \code{overlap_to_interval_len()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedIntervalCover$overlap_to_interval_len(filter, percent_overlap)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedIntervalCover-interval_len_to_percent_overlap"></a>}}
\if{latex}{\out{\hypertarget{method-FixedIntervalCover-interval_len_to_percent_overlap}{}}}
\subsection{Method \code{interval_len_to_percent_overlap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedIntervalCover$interval_len_to_percent_overlap(filter, interval_len)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedIntervalCover-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FixedIntervalCover-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedIntervalCover$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
