% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_rect_cover.R
\docType{class}
\name{FixedRectangularCover-class}
\alias{FixedRectangularCover-class}
\alias{f_rect_cover}
\title{Fixed Rectangular Cover}
\description{
The rectangular cover is multidimensional, two-parameter family of covers. Given the number of
intervals and the overlap percentage between these intervals, this class constructs hyper-rectangular cover of the
given filter space, where individual level sets are distributed uniformly along a rectangular grid.
}
\details{
Fixed rectangular cover reference class.
}
\section{Fields}{

\describe{
\item{\code{num_intervals}}{:= vector of number of bins to cover the Z with (per dimension)}

\item{\code{percent_overlap}}{:= vector of overlap percentages}
}}


\author{
Matt Piekenbrock
}
