% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_imapper.R
\docType{data}
\name{future_imapper_lgl}
\alias{future_imapper_lgl}
\title{future_imapper_lgl}
\format{
An object of class \code{environment} of length 39.
}
\usage{
future_imapper_lgl
}
\description{
future_imapper_lgl
}
\keyword{datasets}
