% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_imapper.R
\name{future_imapper_template}
\alias{future_imapper_template}
\title{Apply a function to each element of a vector, and its index via futures}
\usage{
future_imapper_template(arg)
}
\value{
A vector the same length as .x.

future_imapper_template
}
\description{
These functions work exactly the same as \code{\link[purrr:imap]{purrr::imap()}} functions,
but allow you to map in parallel.
}
