% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_invoke_mapper.R
\name{future_invoke_mapper}
\alias{future_invoke_mapper}
\title{invoke_mapping in Parallel and Ending parallel invoke_mapping \code{furrr} in R}
\usage{
future_invoke_mapper(...)
}
\arguments{
\item{workers}{default parallel workers are detectCores()-2}
}
\description{
invoke_mapping in Parallel and Ending parallel invoke_mapping \code{furrr} in R
}
