% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_invoke_mapper.R
\name{future_invoke_mapper_template}
\alias{future_invoke_mapper_template}
\title{Apply a function to each element of a vector, and its index via futures}
\usage{
future_invoke_mapper_template(...)
}
\arguments{
\item{...}{Additional arguments passed to each function.}
}
\value{
A vector the same length as .x.
future_invoke_mapper_template
}
\description{
These functions work exactly the same as \code{\link[purrr:invoke]{purrr::invoke_map()}} functions,
but allow you to map in parallel.
}
