% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper2.R
\docType{data}
\name{future_mapper2}
\alias{future_mapper2}
\title{Apply a function to each element of a vector, and its index via futures}
\format{
An object of class \code{environment} of length 44.
}
\usage{
future_mapper2
}
\value{
A vector the same length as .x.
}
\description{
These functions work exactly the same as \code{\link[purrr:map2]{purrr::map2()}} functions,
but allow you to map in parallel.
}
\keyword{datasets}
