% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper2.R
\name{future_mapper2_template}
\alias{future_mapper2_template}
\title{Apply a function to each element of a vector, and its index via futures}
\usage{
future_mapper2_template(arg)
}
\value{
A vector the same length as .x.
future_mapper2_template
}
\description{
These functions work exactly the same as \code{\link[purrr:map2]{purrr::map2()}} functions,
but allow you to map in parallel.
}
