% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper.R
\name{future_mapper_template}
\alias{future_mapper_template}
\title{Apply a function to each element of a vector, and its index via futures}
\usage{
future_mapper_template(arg)
}
\value{
A vector the same length as .x.
future_mapper_template
}
\description{
These functions work exactly the same as \code{\link[purrr:imap]{purrr::imap()}} functions,
but allow you to map in parallel.
}
