% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper.R
\docType{data}
\name{future_mapper_walk}
\alias{future_mapper_walk}
\title{future_invoke_mapper_walk}
\format{
An object of class \code{environment} of length 39.
}
\usage{
future_mapper_walk
}
\description{
future_invoke_mapper_walk
}
\keyword{datasets}
