% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_pmapper.R
\docType{data}
\name{future_pmapper}
\alias{future_pmapper}
\title{Apply a function to each element of a vector, and its index via futures}
\format{
An object of class \code{environment} of length 44.
}
\usage{
future_pmapper
}
\value{
A vector the same length as .x.
}
\description{
These functions work exactly the same as \code{\link[purrr:pmap]{purrr::pmap()}} functions,
but allow you to map in parallel.
}
\keyword{datasets}
