% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugeProp.R
\name{gaugeProp}
\alias{gaugeProp}
\title{Gauge proportion}
\usage{
gaugeProp(coords, catchment)
}
\arguments{
\item{coords}{Coordinates of rain gauges set as a polygon shapefile}

\item{catchment}{Catchment shapefile}
}
\value{
Rainfall proportions for each rain gauge
}
\description{
Calculates the the proportion of catchment that the gauge occupies using Thiessen / Voronoi polygons
}
\examples{
library(riskyData)
data(crowle); data(bickley); data(barnhurst); data(hollies); data(ledbury);
data(bettwsYCrwyn); data(bewdCatch)

# Obtain gauge coordinates
gcs <- getCoords(crowle,
                 bickley,
                 barnhurst,
                 hollies,
                 ledbury,
                 bettwsYCrwyn)

gaugeProp(gcs, bewdCatch)
}
