% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballmapper.R
\name{get_ballmapper_data}
\alias{get_ballmapper_data}
\title{Runs ballmapper and returns two dataframes containing node and edge data.}
\usage{
get_ballmapper_data(data, dists, eps)
}
\arguments{
\item{data}{Your input data. Ideally a dataframe.}

\item{dists}{A distance matrix for your data. Can be a \code{dist} object or 2D matrix.}

\item{eps}{A positive real number for your desired ball radius.}
}
\value{
A list of two dataframes, one with node data containing ball size,
datapoints per ball, ball tightness, and one with edge data
containing sources, targets, and weights representing overlap strength.
}
\description{
Runs ballmapper and returns two dataframes containing node and edge data.
}
\examples{
circle.data = data.frame( x= sapply(1:1000, function(x) cos(x)) + rnorm(100, 500, .03),
  y = sapply(1:1000, function(x) sin(x)) + rnorm(100, 0, 0.03))
circle.dist = dist(circle.data)

show(get_ballmapper_data(circle.data, circle.dist, .3))
}
