% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_hub.R
\name{get_bin_vector}
\alias{get_bin_vector}
\title{Recover bins}
\usage{
get_bin_vector(binclust_data)
}
\arguments{
\item{binclust_data}{A list of bins, each containing named vectors whose names are those of data points and whose values are cluster ids.}
}
\value{
A vector of integers equal in length to the number of clusters, whose members identify which bin that cluster belongs to.
}
\description{
Recover bins
}
