% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_hub.R
\name{get_cluster_sizes}
\alias{get_cluster_sizes}
\title{Compute cluster sizes}
\usage{
get_cluster_sizes(binclust_data)
}
\arguments{
\item{binclust_data}{A list of bins, each containing named vectors whose names are those of data points and whose values are cluster ids.}
}
\value{
A vector of integers representing the lengths of the clusters in the input data.
}
\description{
Compute cluster sizes
}
