% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_hub.R
\name{get_clustered_data}
\alias{get_clustered_data}
\title{Get data within a cluster}
\usage{
get_clustered_data(binclust_data)
}
\arguments{
\item{binclust_data}{A list of bins, each containing named vectors whose names are those of data points and whose values are cluster ids}
}
\value{
A list of strings, each a comma separated list of the toString values of the data point names.
}
\description{
Get data within a cluster
}
