% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_factory.R
\name{get_clusters}
\alias{get_clusters}
\title{Initate the clustering process}
\usage{
get_clusters(bins, dists, method)
}
\arguments{
\item{bins}{A list containing "bins" of vectors of names of data points.}

\item{dists}{A distance matrix containing pairwise distances between named data points.}

\item{method}{A string to pass to \link{hclust} to determine clustering method.}
}
\value{
A list containing named vectors (one per bin), whose names are data point names and whose values are cluster labels
}
\description{
This function processes the binned data and global distance matrix to return freshly clustered data.
}
