% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{get_overlaps}
\alias{get_overlaps}
\title{Get cluster overlaps}
\usage{
get_overlaps(binclust_data)
}
\arguments{
\item{binclust_data}{A list of bins, each containing named vectors whose names are those of data points and whose values are cluster ids.}
}
\value{
A named list of edges, whose elements contain the names of clusters in the overlap represented by that edge.
}
\description{
Get cluster overlaps
}
