% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watering_hole.R
\name{get_single_linkage_clusters}
\alias{get_single_linkage_clusters}
\title{Perform single linkage clustering and process dendrograms}
\usage{
get_single_linkage_clusters(dist_mats)
}
\arguments{
\item{dist_mats}{A list of distance matrices to be used for clustering.}
}
\value{
A list containing named vectors (one per dendrogram), whose names are data point names and whose values are cluster labels
}
\description{
Perform single linkage clustering and process dendrograms
}
