% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterer_farm.R
\name{get_tallest_branch_height}
\alias{get_tallest_branch_height}
\title{Find the tallest branch of a dendrogram}
\usage{
get_tallest_branch_height(dend, max_height = max(cophenetic(dend)))
}
\arguments{
\item{dend}{A single dendrogram.}

\item{max_height}{The maximum height of the dendrogram; if this is not provided the last merge height of the input dendrogram will be used, which will make one cluster impossible!}
}
\value{
The height of the tallest branch (longest time between merge heights) of the input dendrogram.
}
\description{
Find the tallest branch of a dendrogram
}
