% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausdorff.R
\name{hausdorff_distance}
\alias{hausdorff_distance}
\title{Hausdorff Distance}
\usage{
hausdorff_distance(m)
}
\arguments{
\item{m}{A \code{MapperRef} object.}
}
\value{
A \code{\link[stats]{dist}} object with \code{Size} equal to the number of vertices.
}
\description{
Computes a smooth approximation to the Hausdorff distance between the vertices of a mapper.
Implements the equation in Section 5 of the original Mapper paper. 
\deqn{
 d_h(X_i, X_j) = \max{( \frac{\sum_y \min_x(d(x, y)) }{n_j}, \frac{\sum_x \min_y(d(x, y)) }{n_i})}
}{
 DH(Xi, Xj) = max\{ \sum min(d(x, y)) / nj, \sum min(d(x, y)) / ni\}
}
where \eqn{x, y} are elements of the clusters \eqn{C_i, C_j}{Ci, Cj} respectively, and \eqn{X_i, X_j}{Xi, Xj} are the clusters corresponding 
vertices in the mapper. See the reference below for more details. \cr
\cr
\strong{WARNING:} This function may be very computationally expensive.
}
\details{
Currently, this requires the \pkg{RANN} package to be installed, and only considers euclidean distance.
}
\references{
Singh, Gurjeet, Facundo Memoli, and Gunnar E. Carlsson. "Topological methods for the analysis of high dimensional data sets and 3d object recognition." SPBG. 2007.
}
