% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_imapper.R
\name{imapper_dfc}
\alias{imapper_dfc}
\title{imapper_dfc}
\usage{
imapper_dfc(workers = (parallelly::availableCores() - 2), ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
imapper_dfc
}
