% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_imapper.R
\name{imapper_dfr}
\alias{imapper_dfr}
\title{imapper_dfr}
\usage{
imapper_dfr(workers = NULL, ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
imapper_dfr
}
