% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_imapper.R
\name{imapper_int}
\alias{imapper_int}
\title{imapper_int}
\usage{
imapper_int(workers = parallel::detectCores() - 2, ...)
}
\arguments{
\item{workers}{default value is the \verb{total cores - 2}}
}
\description{
imapper_int
}
