% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_invoke_mapper.R
\name{invoke_mapper}
\alias{invoke_mapper}
\title{Apply a function to each element of a vector, and its index via futures}
\usage{
invoke_mapper(...)
}
\arguments{
\item{...}{Additional arguments passed to each function.}
}
\value{
parameters
}
\description{
These functions work exactly the same as \code{\link[purrr:invoke]{purrr::invoke_map()}} functions,
but allow you to map in parallel.
}
