% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_invoke_mapper.R
\name{invoke_mapper_chr}
\alias{invoke_mapper_chr}
\title{invoke_mapper_chr}
\usage{
invoke_mapper_chr(workers = (parallelly::availableCores() - 2), ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
invoke_mapper_chr
}
