% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_invoke_mapper.R
\name{invoke_mapper_lgl}
\alias{invoke_mapper_lgl}
\title{invoke_mapper_lgl}
\usage{
invoke_mapper_lgl(workers = NULL, ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
invoke_mapper_lgl
}
