% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilting_hut.R
\name{is_in_ball}
\alias{is_in_ball}
\title{Get a tester function for a ball.}
\usage{
is_in_ball(ball)
}
\arguments{
\item{ball}{A list of data points.}
}
\value{
A function that eats a data point and returns TRUE or FALSE depending if the point is in the ball or not.
}
\description{
Get a tester function for a ball.
}
