% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks.R
\name{landmarks}
\alias{landmarks}
\title{landmarks}
\usage{
landmarks(
  x,
  n,
  dist_method = "euclidean",
  seed_index = 1,
  shuffle_data = FALSE
)
}
\arguments{
\item{x}{a data matrix.}

\item{n}{the number of landmarks requested.}

\item{dist_method}{the distance metric to use. Any distance measure in the \code{proxy} package is supported.}

\item{seed_index}{the first landmark to seed the algorithm.}

\item{shuffle_data}{whether to first randomly shuffle the data.}
}
\description{
Computes a set of \code{n} landmarks of a point cloud.
}
\details{
This function uses the greedy maxmin procedure to produce a set of \code{n} evenly spaced landmark points in the data set 
\code{x}. Maxmin is a simple greedy algorithm that is relatively efficient, but it has a tendency to pick out extremal points. 
If the distance metric is euclidean, an efficient Rcpp implementation is used. If another metric is requested, 
the algorithm is performed in R.
}
\references{
De Silva, Vin, and Gunnar E. Carlsson. "Topological estimation using witness complexes." SPBG 4 (2004): 157-166.
}
