% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{linkedTable}
\alias{linkedTable}
\title{linkedTableOutput - shiny server module}
\usage{
linkedTable(input, output, session, dt, config = NULL, options = NULL)
}
\description{
Utility function to make a datatables object with nice default extensions.
}
