% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCatchment.R
\name{loadCatchment}
\alias{loadCatchment}
\title{Import catchment shapefiles suitable for flode}
\usage{
loadCatchment(filepath = NULL)
}
\arguments{
\item{filepath}{link to the shapefile of interest}
}
\value{
A catchment polygon set to the WGS 84 projection
}
\description{
Import catchment shapefiles suitable for flode
}
\examples{
## Do not run
# loadCatchment(filepath = "c://abcd/efg/hi/jkl.shp")
}
