% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkroom.R
\name{make_bins}
\alias{make_bins}
\title{Create bins of data from a 1D filter function}
\usage{
make_bins(data, filtered_data, bin_ends)
}
\arguments{
\item{data}{A dataframe.}

\item{filtered_data}{A single column of the input dataframe.}

\item{bin_ends}{A 2D numeric array of input ranges, whose first column contains left endpoints and whose second column contains right endpoints.}
}
\value{
A list of "bins" containing vectors of names of datapoints.
}
\description{
This function identifies datapoints whose "filter" value falls within a specified set of numeric ranges.
}
