% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkroom.R
\name{make_one_bin}
\alias{make_one_bin}
\title{Create a single bin of data}
\usage{
make_one_bin(bin_left, bin_right, data, filtered_data)
}
\arguments{
\item{bin_left}{The minimum value to be filtered for.}

\item{bin_right}{The maximum value to be filtered for.}

\item{data}{A dataframe.}

\item{filtered_data}{A single column of the input dataframe.}
}
\value{
A vector of datapoint names which fall within the filter range.
}
\description{
This function identifies datapoints whose "filter" value falls within a numeric range.
}
