% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_imapper.R, R/future_invoke_mapper.R,
%   R/future_mapper.R, R/future_mapper2.R, R/future_pmapper.R
\name{map_func}
\alias{map_func}
\title{MAPping in Parallel and Ending parallel mapping \code{furrr} in R}
\arguments{
\item{workers}{default parallel workers are detectCores()-2}
}
\description{
MAPping in Parallel and Ending parallel mapping \code{furrr} in R

MAPping in Parallel and Ending parallel mapping \code{furrr} in R

MAPping in Parallel and Ending parallel mapping \code{furrr} in R

MAPping in Parallel and Ending parallel mapping \code{furrr} in R

MAPping in Parallel and Ending parallel mapping \code{furrr} in R
}
