% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapper-package.r
\docType{package}
\name{Mapper}
\alias{Mapper}
\alias{Mapper-package}
\title{Mapper Package}
\description{
Provides an R/Rcpp implementation of Mapper and its related tools. 
The main documentation for the package can be found here: \url{https://peekxc.github.io/mapper/}.
}
\author{
Matt Piekenbrock
}
