% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper2.R
\name{mapper2}
\alias{mapper2}
\title{Apply a function to each element of a vector, and its index via futures}
\usage{
mapper2(..., workers = NULL)
}
\arguments{
\item{...}{Additional arguments passed on to the mapped function.

We now generally recommend against using \code{...} to pass additional
(constant) arguments to \code{.f}. Instead use a shorthand anonymous function:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Instead of
x |> map(f, 1, 2, collapse = ",")
# do:
x |> map(\\(x) f(x, 1, 2, collapse = ","))
}\if{html}{\out{</div>}}

This makes it easier to understand which arguments belong to which
function and will tend to yield better error messages.}

\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\value{
parameters
}
\description{
These functions work exactly the same as \code{\link[purrr:map2]{purrr::map2()}} functions,
but allow you to map in parallel.
}
