% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper2.R
\name{mapper2_chr}
\alias{mapper2_chr}
\title{mapper2_chr}
\usage{
mapper2_chr(workers = NULL, ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
mapper2_chr
}
