% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper2.R
\name{mapper2_dbl}
\alias{mapper2_dbl}
\title{mapper2_dbl}
\usage{
mapper2_dbl(workers = NULL, ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
mapper2_dbl
}
