% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper2.R
\name{mapper2_dfr}
\alias{mapper2_dfr}
\title{mapper2_dfr}
\usage{
mapper2_dfr(workers = (parallelly::availableCores() - 2), ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
mapper2_dfr
}
