% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper2.R
\name{mapper2_int}
\alias{mapper2_int}
\title{mapper2_int}
\usage{
mapper2_int(workers = parallel::detectCores() - 2, ...)
}
\arguments{
\item{workers}{default value is the \verb{total cores - 2}}
}
\description{
mapper2_int
}
