% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper2.R
\name{mapper2_lgl}
\alias{mapper2_lgl}
\title{mapper2_lgl}
\usage{
mapper2_lgl(workers = parallel::detectCores() - 2, ...)
}
\arguments{
\item{workers}{default value is the \verb{total cores - 2}}
}
\description{
mapper2_lgl
}
