% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper2.R
\name{mapper2_walk2}
\alias{mapper2_walk2}
\title{mapper2_walk2}
\usage{
mapper2_walk2(workers = (parallelly::availableCores() - 2), ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
mapper2_walk2
}
