% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper.R
\name{mapper_chr}
\alias{mapper_chr}
\title{mapper_chr}
\usage{
mapper_chr(workers = parallel::detectCores() - 2, ...)
}
\arguments{
\item{workers}{default value is the \verb{total cores - 2}}
}
\description{
mapper_chr
}
