% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper.R
\name{mapper_dbl}
\alias{mapper_dbl}
\title{mapper_dbl}
\usage{
mapper_dbl(workers = (parallelly::availableCores() - 2), ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
mapper_dbl
}
