% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper.R
\name{mapper_dfc}
\alias{mapper_dfc}
\title{mapper_dfc}
\usage{
mapper_dfc(..., workers = NULL)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
mapper_dfc
}
