% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper.R
\name{mapper_dfr}
\alias{mapper_dfr}
\title{mapper_dfr}
\usage{
mapper_dfr(workers = parallel::detectCores() - 2, ...)
}
\arguments{
\item{workers}{default value is the \verb{total cores - 2}}
}
\description{
mapper_dfr
}
