% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper.R
\name{mapper_lgl}
\alias{mapper_lgl}
\title{mapper_lgl}
\usage{
mapper_lgl(workers = parallel::detectCores() - 2, ...)
}
\arguments{
\item{workers}{default value is the \verb{total cores - 2}}
}
\description{
mapper_lgl
}
