% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{MapperData}
\alias{MapperData}
\title{Shiny data}
\usage{
MapperData(input, output, session)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}
}
\value{
reactive expression that returns a function containing the data set.
}
\description{
Shiny data
}
