% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{MapperFilterInput}
\alias{MapperFilterInput}
\title{Mapper filter shiny UI module}
\usage{
MapperFilterInput()
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{datasets}{optional named list of data sets. See details.}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements.
}
\description{
Creates the UI to select a data set for use with \code{ShinyMapper}.
}
\details{
if \code{datasets} is supplied, it must be a named list, where each name corresponds
to the data set name and each element is either a matrix or a function that accepts as input 
a set of indices and returns either a data matrix or a dist object. See \code{\link{MapperRef}} for details.
}
