% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{mapperOptions}
\alias{mapperOptions}
\title{Mapper input options - shiny server module}
\usage{
mapperOptions(input, output, session, mapper)
}
\description{
Mapper input options - shiny server module
}
