% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{MapperRef.plot}
\alias{MapperRef.plot}
\title{Plot mapper with base graphics.}
\arguments{
\item{vertex_min}{minimum vertex cex. Default is 2.}

\item{vertex_max}{maximum vertex cex. Default is 4.}
}
\description{
This function plots the mapper using base \code{\link[graphics]{graphics}}. 
The logic to perform the actual plotting is handled by \code{\link[simplextree]{plot.simplextree}}. 
This function assumes some traditional visualization parameters often used to visualize the mapper 
complex.
}
\seealso{
Other mapper visualization functions: \code{\link{as_igraph}},
  \code{\link{as_pixiplex}}
}
\concept{mapper visualization functions}
