% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapper-package.r
\docType{data}
\name{MapperVis}
\alias{MapperVis}
\title{MapperVis}
\format{An object of class \code{module} (inherits from \code{list}) of length 6.}
\usage{
MapperVis
}
\description{
MapperVis
}
\keyword{datasets}
