% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{mapperVisOutput}
\alias{mapperVisOutput}
\title{mapperVis - shiny UI module}
\usage{
mapperVisOutput(id, output_type = c("grapher"))
}
\description{
Module for configuring the visualization options for Mapper.
}
