\name{mclustmappeR}
\alias{mclustmappeR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Clustering via mclust
}
\description{
Computes a Normal mixture model based clustering of the given data.
}
\usage{
mclustmappeR
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{Data matrix}
}
\details{
Pass as an argument to mappeR. Additional arguments via can be passed to mclust via ..., but a function will need to be defined first and then passed to mappeR.
}
\value{A vector of cluster memberships.}
\references{}
\author{
Matthew Arnold
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line