% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{next_triangular}
\alias{next_triangular}
\title{Find which triangular number you're on}
\usage{
next_triangular(x)
}
\arguments{
\item{x}{A positive integer.}
}
\value{
The index of the next greatest or equal triangular number to \eqn{x}.
}
\description{
Find which triangular number you're on
}
