% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{nodeHighlight}
\alias{nodeHighlight}
\title{nodeHighlight shiny server module}
\usage{
nodeHighlight(mapper_vis, reset)
}
\description{
Server-only function to 'highlight' nodes. This does the equivalent of changing the nodes colors, 
but adjusts low values to have lower opacity, and registers a reactive expression to to reset when 
no data points are selected. Takes as input a boolean vector indicated which _points_ should be highlighted.
}
