% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{nodeSizingInput}
\alias{nodeSizingInput}
\title{Node sizing input/UI module}
\usage{
nodeSizingInput(id, node_min_rng = c(1, 25), node_max_rng = c(1, 50),
  node_scaling = c("linear", "logarithmic"))
}
\description{
Node sizing input/UI module
}
