% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noisy_circle}
\alias{noisy_circle}
\title{1,500 noisy points sampled on the perimeter of a circle.}
\format{
A two dimensional of points.
}
\source{
Generated.
}
\usage{
noisy_circle
}
\description{
The circle has radius ~ 2.1 and is based on example 3.2 of [1].
}
\examples{
## Example code to generate noisy points around the perimeter of a circle 
n <- 1500
t <- 2*pi*runif(n)
r <- runif(n, min = 2, max = 2.1)
noisy_circle <- cbind(r*cos(t), r*sin(t))
}
\references{
Singh, Gurjeet, Facundo Mémoli, and Gunnar E. Carlsson. "Topological methods for the analysis of high dimensional data sets and 3d object recognition." SPBG. 2007.
}
\keyword{datasets}
