% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{plot.Mapper}
\alias{plot.Mapper}
\title{S3 method for default plotting}
\usage{
\method{plot}{Mapper}(x, ...)
}
\arguments{
\item{x}{Mapper object.}

\item{...}{unused.}
}
\description{
S3 method for default plotting
}
