% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{plot}
\alias{plot}
\title{Plot the annotated simplicial complex}
\description{
Renders a plot of the constructed skeleton with constructed
  annotation (if any).
}
\details{
The simplicial complex produced by the mapper construction can be
  plotted directly using the \code{\link{plot()}} method of
  \link[simplextree:simplextree]{simplex tree}. This \code{MapperRef} method
  is a shortcut to this method that additionally encodes the number of
  observations in each simplex and their staining variable values as sizes
  and colors.
}
