% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_pmapper.R
\name{pmapper_dbl}
\alias{pmapper_dbl}
\title{pmapper_dbl}
\usage{
pmapper_dbl(workers = (parallelly::availableCores() - 2), ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
pmapper_dbl
}
