% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_pmapper.R
\name{pmapper_dfc}
\alias{pmapper_dfc}
\title{pmapper_dfc}
\usage{
pmapper_dfc(workers = parallel::detectCores() - 2, ...)
}
\arguments{
\item{workers}{default value is the \verb{total cores - 2}}
}
\description{
pmapper_dfc
}
