% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_pmapper.R
\name{pmapper_dfr}
\alias{pmapper_dfr}
\title{pmapper_dfr}
\usage{
pmapper_dfr(..., workers = NULL)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
pmapper_dfr
}
