% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_pmapper.R
\name{pmapper_int}
\alias{pmapper_int}
\title{pmapper_int}
\usage{
pmapper_int(workers = NULL, ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
pmapper_int
}
