% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_pmapper.R
\name{pmapper_lgl}
\alias{pmapper_lgl}
\title{pmapper_lgl}
\usage{
pmapper_lgl(workers = NULL, ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
pmapper_lgl
}
