% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_pmapper.R
\name{pmapper_pwalk}
\alias{pmapper_pwalk}
\title{pmapper_pwalk}
\usage{
pmapper_pwalk(workers = parallel::detectCores() - 2, ...)
}
\arguments{
\item{workers}{default value is the \verb{total cores - 2}}
}
\description{
pmapper_pwalk
}
