% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{print.Mapper}
\alias{print.Mapper}
\title{S3 method for default printing}
\usage{
\method{print}{Mapper}(x, ...)
}
\arguments{
\item{x}{Mapper object.}

\item{...}{unused.}
}
\description{
S3 method for default printing
}
