% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestrainedIntervalCover.R
\docType{class}
\name{RestrainedIntervalCover}
\alias{RestrainedIntervalCover}
\title{Restrained Interval Cover}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RestrainedIntervalCover
}
\description{
The interval cover is multidimensional, two-parameter family of covers. Given the number of
intervals and the overlap percentage between these intervals, this class constructs hyper-rectangular cover of the
given filter space, where individual level sets are distributed uniformly along a rectangular grid.
}
\section{Fields}{

\describe{
\item{\code{number_intervals}}{vector of number of bins to cover the Z with (per dimension)}

\item{\code{percent_overlap}}{vector of overlap percentages}
}}

\author{
Matt Piekenbrock
}
\keyword{datasets}
