% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_rect_cover.R
\docType{class}
\name{RestrainedRectangularCover-class}
\alias{RestrainedRectangularCover-class}
\alias{r_rect_cover}
\title{Restrained Rectangular Cover}
\description{
The rectangular cover is multidimensional, two-parameter family of covers. Given the number of
intervals and the overlap percentage between these intervals, this class constructs hyperrectangular cover of the
given filter space, where individual level sets are distributed uniformly along a rectangular grid.
}
\details{
Restrained rectangular cover reference class - part of Reference Class (R5) implementation for the Mapper package in R
}
\section{Fields}{

\describe{
\item{\code{num_intervals}}{:= vector of number of bins to cover the Z with (per dimension)}

\item{\code{percent_overlap}}{:= vector of overlap percentages}
}}


\author{
Matt Piekenbrock
}
