% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watering_hole.R
\name{run_slink}
\alias{run_slink}
\title{Perform single linkage clustering}
\usage{
run_slink(dist)
}
\arguments{
\item{dist}{A distance matrix.}
}
\value{
A dendrogram generated by \code{fastcluster}.
}
\description{
Perform single linkage clustering
}
