% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_tree.R
\name{segment_tree}
\alias{segment_tree}
\title{Segment Tree}
\usage{
segment_tree(intervals)
}
\arguments{
\item{intervals}{an (n x 2) matrix of intervals.}
}
\value{
A queryable segment tree.
}
\description{
Fast segment tree class exposed as an Rcpp Module.
}
\details{
A segment tree is rooted-augment binary search tree useful for querying information across a fixed set of intervals.
This segment tree class is specialized to handle interval queries which return indices of points that intersect the given query interval.
The query interval must be composed using endpoints which were used to construct the tree.
}
\references{
Mehta, Dinesh P., and Sartaj Sahni. "Chapter 18: Interval, Segment, Range, and Priority Search Trees." Handbook of data structures and applications. Chapman and Hall/CRC, 2004.
}
