% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{ShinyMapper}
\alias{ShinyMapper}
\title{Shiny Mapper module server function}
\usage{
ShinyMapper(input, output, session, dataset, cover, M = NULL)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}

\item{dataset}{reactive expression that returns the data set.}

\item{cover}{reactive expression that returns the cover.}

\item{M}{optional MapperRef instance (non-reactive). See details.}
}
\value{
rective expression that constructs the Mapper.
}
\description{
module server function for shiny mapper.
}
\details{
If \code{M} is supplied, the parameters of the current Mapper are copied to the inputs, and that instance 
is then updated using the controls. Otherwise, a new \code{\link{MapperRef}} instance is constructed and returned from 
this method, parameterized by the associated inputs.
}
