% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{ShinyMapperInput}
\alias{ShinyMapperInput}
\title{Shiny Mapper module UI function}
\usage{
ShinyMapperInput(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Creates UI elements that parameterize the Mapper construction for Shiny apps.
}
