% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{simplex_tree}
\alias{simplex_tree}
\title{Simplex Tree}
\usage{
simplex_tree()
}
\value{
A queryable simplex tree.
}
\description{
Simplex tree class exposed as an Rcpp Module.
}
\details{
A simplex tree is a trie specialized for storing simplicial complexes.
}
\references{
Boissonnat, Jean-Daniel, and Clement Maria. "The simplex tree: An efficient data structure for general simplicial complexes." Algorithmica 70.3 (2014): 406-427.
}
