% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{simplicial_complex}
\alias{simplicial_complex}
\title{Simplicial Complex}
\description{
The relational information of the Mapper construction.
}
\details{
The primary output of the Mapper method is a simplicial complex. With \code{\link{MapperRef}} objects, 
the simplicial complex is stored as a \code{\link[simplextree]{simplextree}}. 
\cr 
The underlying complex does not need to be modified by the user, i.e. is completely maintained by \code{\link{MapperRef}} methods 
(e.g. \code{\link{construct_nerve}}, \code{\link{construct_k_skeleton}}, etc.).
}
