% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquareDualTessellationCover.R
\docType{class}
\name{SquareDualTessellationCover}
\alias{SquareDualTessellationCover}
\title{Square Dual Tesselation Cover}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SquareDualTessellationCover
}
\description{
[R6] class representing a square dual tessellation cover.
}
\details{
The regular square tessellation \eqn{T} of \eqn{Z} (the plane, or a
path-connected region of it; Schlafli symbol \eqn{{4,4}}) consists of
congruent squares arranged as tiles, with shared vertices and edges and empty
overlap. In order to completely partition \eqn{Z}, each square is considered
closed along half of its perimeter and open along the other half. (**This
remains to be carefully implemented.**) The dual \eqn{T'} of \eqn{T} is
obtained by taking vertices at the centroids of the tiles of \eqn{T} and
edges perpendicular to those of \eqn{T}. The union \eqn{C} of \eqn{T} and
\eqn{T'} is a _square dual tessellation cover_.

A square dual tessellation cover \eqn{C} is the union of two
`[FixedIntervalCover]`s, for a 2-dimensional lens and with the same uniform
`number_intervals` and  `percent_overlap = 0`, each offset from the other by
half of the side length along each dimension. \eqn{C} has constant "depth",
in the sense that every point in \eqn{Z} is a member of exactly two sets in
\eqn{C}. A cover of depth 2 is only sufficient to recover 1-dimensional
features of a point cloud. Every non-empty intersection of cover sets is
congruent up to boundary.

`SquareDualTessellationCover` takes the centroid of \eqn{f(X)} to be the
centroid of the overlap between a pair of tiles in \eqn{T} and in \eqn{T'}.
Thus, the tessellations have rotation symmetry and the same number of sets
(before dropping empty sets). When `width` is sufficiently large (twice the
range of the data), the total number of tiles is minimized at 2.
}
\section{Fields}{

\describe{
\item{\code{width}}{A positive number. The side length of each square in the cover,
as a fraction of the larger coordinate range of the lensed data.}
}}

\seealso{
Other cover: \code{\link{FixedIntervalCover}},
  \code{\link{IntervalDualTessellationCover}},
  \code{\link{TriangleDualTessellationCover}}
}
\author{
Jason Cory Brunson
}
\concept{cover}
\keyword{datasets}
