% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace_cover.R
\docType{class}
\name{SubspaceCover}
\alias{SubspaceCover}
\title{Subspace Cover}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SubspaceCover
}
\description{
The subspace cover is a special type of cover to represent subspace topologies. 
That is, if \eqn{X} is a topological space with topology \eqn{\tau}, and \eqn{Y} is a 
subset of \eqn{X}, the collection
\deqn{\delta = { Y ⋂ U | U ∈ \tau }}
is a topology on \eqn{Y}, called the \emph{subspace topology}. The open sets of the subspace 
topology consist of the intersections between the open sets of \eqn{X} with \eqn{Y}.\cr 
\cr 
The subspace cover represents subspaces by a vector of integer indices between \eqn{[1, n]}, where 
\eqn{n} is the number points in \eqn{X}. This specific class allows the creation of two types of subspaces: 
1)
}
\section{Fields}{

\describe{
\item{\code{number_intervals}}{:= vector of number of bins to cover the Z with (per dimension)}

\item{\code{percent_overlap}}{:= vector of overlap percentages}
}}

\author{
Matt Piekenbrock
}
\keyword{datasets}
