% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeSun.R
\name{teeSun}
\alias{teeSun}
\title{teeSun}
\usage{
teeSun(gaugeCoords, catchment = NULL)
}
\arguments{
\item{gaugeCoords}{SF file with point data within.}

\item{catchment}{Set as null, optional bounding box polygon derived from
catchment polygon which is useful for large catchments}
}
\value{
A polygon shapefile with Thiessen polygons
}
\description{
Creates a Thiessen/Voronoi polygon shapefile.
}
\examples{
library(riskyData)
data(crowle); data(bickley); data(barnhurst); data(hollies); data(ledbury);
data(bettwsYCrwyn); data(bewdCatch)

# Obtain gauge coordinates
gcs <- getCoords(crowle,
                 bickley,
                 barnhurst,
                 hollies,
                 ledbury,
                 bettwsYCrwyn)

bewdTeeSun <- teeSun(gaugeCoords = gcs, catchment = bewdCatch)

leaflet(bewdTeeSun) \%>\%
  addTiles() \%>\%
  addPolygons()
}
