% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teesunPlot.R
\name{teeSunPlot}
\alias{teeSunPlot}
\title{teesunPlot}
\usage{
teeSunPlot(
  gaugeCoords,
  intersectPoly,
  polyCol = c("green", "yellow", "red"),
  pointCol = "orangered"
)
}
\arguments{
\item{gaugeCoords}{Coordinates of rain gauges set as a polygon shapefile}

\item{intersectPoly}{Catchment shapefile}
}
\value{
An interactive Thiessen polygon map
}
\description{
Creates a map of a catchment with Thiessen / Voronoi polyggons overlayed
}
\examples{
library(riskyData)
data(crowle); data(bickley); data(barnhurst); data(hollies); data(ledbury);
data(bettwsYCrwyn); data(bewdCatch)

# Obtain gauge coordinates
gcs <- getCoords(crowle,
                 bickley,
                 barnhurst,
                 hollies,
                 ledbury,
                 bettwsYCrwyn)

bewdTeeSun <- teeSun(gaugeCoords = gcs, catchment = bewdCatch)
int <- intersectPoly(coords = gcs,
                     voronoi = bewdTeeSun,
                     catchment = bewdCatch)
teeSunPlot <- teeSunPlot(guageCoords = gcs,
                         intersectPoly = int)
}
