% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleDualTessellationCover.R
\docType{class}
\name{TriangleDualTessellationCover}
\alias{TriangleDualTessellationCover}
\title{Triangle Dual Tesselation Cover}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TriangleDualTessellationCover
}
\description{
[R6] class representing a triangle--hexagon dual tessellation
  cover.
}
\details{
The regular triangle tessellation \eqn{T} of \eqn{Z} (the plane, or a
path-connected region of it; Schlafli symbol \eqn{{3,6}}) consists of
congruent equiangular triangles arranged as tiles, with shared vertices and
edges and empty overlap. In order to completely partition \eqn{Z}, each
triangle is considered closed along part of its perimeter and open along the
remainder. (**This remains to be carefully implemented.**) The dual \eqn{T'}
of \eqn{T} (the regular hexagonal tessellation; Schlafli symbol \eqn{6,3}) is
obtained by taking vertices at the centroids of the tiles of \eqn{T} and
edges perpendicular to those of \eqn{T}. The union \eqn{C} of \eqn{T} and
\eqn{T'} is a _triangle--hexagon dual tessellation cover_.

A triangle--hexagon dual tessellation cover \eqn{C} has constant "depth", in
the sense that every point in \eqn{Z} is a member of exactly two sets in
\eqn{C}. A cover of depth 2 is only sufficient to recover 1-dimensional
features of a point cloud. Every non-empty intersection of cover sets is
congruent up to boundary.

`TriangleDualTessellationCover` takes the centroid of \eqn{f(X)} to be the
centroid of the overlap between a pair of tiles in \eqn{T} and in \eqn{T'}.
When `width` is sufficiently large, the total number of tiles is minimized at
6 (though some may be empty).
}
\section{Fields}{

\describe{
\item{\code{width}}{A positive number. The side length of each triangle tile and the
side-to-side length of each hexagon tile, as a fraction of the larger
coordinate range of the lensed data.}
}}

\seealso{
Other cover: \code{\link{FixedIntervalCover}},
  \code{\link{IntervalDualTessellationCover}},
  \code{\link{SquareDualTessellationCover}}
}
\author{
Jason Cory Brunson
}
\concept{cover}
\keyword{datasets}
