% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{use_cover}
\alias{use_cover}
\title{Selects one of the available covering methods to use.}
\arguments{
\item{cover}{Either a pre-configured cover, or name of the cover to use.}

\item{...}{Additional parameter values to pass to the covering generators initialize method.}
}
\description{
Convenience method to select, parameterize, and construct a cover and associate it with the calling objects \code{cover} field.
}
\details{
The \code{cover} parameter can be any \code{\link{CoverRef}}-derived instance (e.g. a \code{\link{FixedIntervalCover}}), 
or one of the cover typenames listed in the \code{\link{covers_available}()}. 
If a typename is given, the cover is automatically constructed before being assigned to the \code{$cover} field.
}
