% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{use_data}
\alias{use_data}
\title{Sets the data}
\arguments{
\item{data}{either a data set name, a matrix, or a function. See details.}

\item{...}{additional parameters to pass to the filter function.}
}
\description{
Sets the data matrix to associate with the mapper.
}
\details{
This function sets the data for the Mapper to work on. If \code{data} is a string, 
it must be one of the (illustrative) data sets listed in \code{data(package="Mapper")}. 
Otherwise, \code{data} must be either a matrix of coordinate values, a function that returns a matrix of 
coordinate values.
}
