% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{use_distance_measure}
\alias{use_distance_measure}
\title{Assign a distance measure}
\description{
Assigns a distance measure to the \code{\link{MapperRef}} instance to use in the clustering algorithm.
}
\section{Usage}{

\preformatted{ $use_distance_measure(measure, ...) }
}

\section{Arguments}{
 
\describe{
  \item{\code{measure}}{The distance measure to use (string).}
  \item{\code{...}}{Extra parameters passed to the distance function. See details. }
}
}

\section{Details}{
 Unless the \code{\link[Mapper:use_clustering_algorithm]{clustering_algorithm}} has been replaced by the user, 
by default, Mapper requires a notion of distance between objects to be defined in creating the vertices of the construction. 

The distance function is determined based on the supplied \code{measure}. \code{measure} must be one of:
\preformatted{["euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"]}
or, if the \pkg{proxy} and \code{\link[parallelDist]{parallelDist}} packages are installed, any name in \code{proxy::pr_DB$get_entry_names()}.\cr
\cr
Additional parameters passed via \code{...} are passed to either \code{\link[stats]{dist}} (or 
\code{\link[parallelDist]{parallelDist}} if installed).
}

\section{Value}{

The mapper instance, with the measure field assigned.
}

\examples{
data(noisy_circle)
m <- MapperRef$new(noisy_circle)
m$use_filter(noisy_circle[,1])
m$use_cover("fixed interval", number_intervals = 5, percent_overlap = 25)

## Constructs clusters with euclidean metric (default)
m$use_distance_measure("euclidean")
m$construct_pullback() 

## Constructs clusters with p-norm (p = 1)
m$use_distance_measure("Minkowski", p = 1L)
m$construct_pullback() 

\dontrun{
## To see list of available measures, use:
proxy::pr_DB$get_entry_names()
}
}
\seealso{
\code{\link[parallelDist]{parDist}} \code{\link[proxy]{pr_DB}}
}
