% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{use_filter}
\alias{use_filter}
\title{Sets the filter}
\arguments{
\item{filter}{either the filter name to use, or a function. See details.}

\item{...}{additional parameters to pass to the filter function.}
}
\description{
Sets the map, or \emph{filter}, to associate with the instance
}
\details{
\code{filter} one of the pre-configured named filters returning matrix-coercible set of coordinate values.
}
