% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{use_layout}
\alias{use_layout}
\title{Choose a layout algorithm}
\arguments{
\item{layout}{The type of layout to create or use. Either a valid string, a
function, or a matrix of coordinates. See details.}

\item{...}{extra parameters passed to a layout generator. See details.}
}
\description{
Assigns a layout algorithm to the \code{\link{MapperRef}}
  instance.
}
\details{
(Add details.)
}
\section{Usage}{
 \preformatted{ $use_layout(layout = "auto, ...) }
}

