% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{use_palette}
\alias{use_palette}
\title{Assign a color palette}
\arguments{
\item{palette}{the name of the palette to use to construct a color scheme.
See details.}

\item{stain}{a vector of length \code{nrow(self$data())} encoding the
variable by which to color simplices. If \code{NULL}, the filter is used.}

\item{...}{extra parameters passed to the palette generator. See details.}
}
\description{
Assigns a color palette to the \code{\link{MapperRef}} instance
  to use when constructing a color scheme.
}
\details{
\code{$use_palette()} does not assign colors to the observations or
  simplices directly but uses a color palette to encode the values of an
  observation-level staining variable. The color of a simplex is determined
  by aggregating the values of this variable at its constituent observations
  (without duplication). The aggregations are mean values if the variable is
  continuous and plurality values if the variable is discrete.

  \code{palette} must be the name of a palette listed in
  \code{palettes_available} and must be compatible with the variable to be
  encoded. (Numeric and logical variables can use continuous palettes;
  character, factor, and logical variables can use discrete palettes.) If
  \code{stain} is provided, then it is verified to have length
  \code{nrow(self$data())} and used as this variable; otherwise, (the first
  column of) the filter \code{self$filter()} is used.

  If the package needed to compute the palette is not installed, a prompt is
  given asking the user whether they would like to install it.
}
\section{Usage}{

\preformatted{
  $use_palette(palette = "rainbow", stain = NULL, ...)
}
}

