% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{vis_options}
\alias{vis_options}
\title{Computes mapper visualization options.}
\arguments{
\item{f}{numeric function to highlight the simplices with. Defaults to mean filter value. See details.}

\item{vertex_scale}{function to scale vertex sizes (either linear or logarithmic). Defaults to linear.}

\item{vertex_min}{minimum vertex size. Default is 10.}

\item{vertex_min}{maximum vertex size. Default is 15.}

\item{embed}{whether to embed the mapper in a metric space with multidimensional scaling. Defaults to false.}

\item{col_pal}{color palette to color the simplices by. Defaults to rainbow.}

\item{...}{extra parameters passed to \code{\link[simplextree]{plot.simplextree}}.}
}
\description{
Computes mapper visualization options.
}
