% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wvs_us_wave6}
\alias{wvs_us_wave6}
\title{Subset of World Values Survey}
\format{
A 28-dimensional data.frame of questionnaire responses, unnormalized.
}
\source{
\url{http://www.worldvaluessurvey.org/WVSDocumentationWV6.jsp}
}
\usage{
wvs_us_wave6
}
\description{
This data set contains a small subset of Wave 6 of the World Values Survey. The subset is 
specifically only responses collected in the U.S., and contains a fraction of the dimensions, 
which were renamed to more semantically clearer names.
}
\details{
The dimensions correspond the following question codes (in order):
V4, V5, V7, V8, V9, V10, V11, V23, V24, V25, V29, V30, V32, V55, V56, V57, V58, V59, V60, 
V61, V67, V68, V69, V84, V95, V97, V98, V101

Please refer to the WV6 codebook for more details on the questions and responses.
}
\keyword{datasets}
