% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_cover.R
\docType{class}
\name{AdaptiveCover}
\alias{AdaptiveCover}
\title{Adaptive Cover}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AdaptiveCover
}
\description{
This class provides cover which 'adapts' the lengths of the open sets to the density of 
the filter values. The cutoff points are chosen with pre-selected quantile method, see \code{?quantile}
for more details on the calculation. Note that the percentage that each bin overlaps is calculated before 
transforming the cutoff points, so the interpretation of the bins as overlapping by \code{percent_overlap}%
isn't quite accurate.
}
\section{Fields}{

\describe{
\item{\code{number_intervals}}{:= vector of number of bins to cover the filter space with (per dimension)}

\item{\code{percent_overlap}}{:= vector of overlap percentages}

\item{\code{quantile_method}}{:= the method to use to compute the quantiles}
}}

\author{
Matt Piekenbrock
}
\keyword{datasets}
