% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{as_igraph}
\alias{as_igraph}
\title{Exports Mapper as an 'igraph' object.}
\arguments{
\item{vertex_scale}{scaling function for the vertex sizes.}

\item{vertex_limits}{minimum and maximum vertex size.}

\item{edge_scale}{scaling function for the edge widths.}

\item{edge_limits}{minimum and maximum edge widths.}
}
\value{
an igraph object.
}
\description{
Exports the 1-skeleton to a graph using the igraph library.
}
\details{
This method converts the 1-skeleton of the Mapper to an igraph
  object and calculates visual properties from the \code{$annotation} (which
  is constructed if necessary): the vertex attributes "color", "size", and
  "label" and the edge attributes "color" and "width". The vertex and edge
  sizes are scaled according to the \code{*_scale} parameters and bounded by
  the \code{*_limits} parameters.
}
