% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{as_igraph}
\alias{as_igraph}
\title{Exports Mapper as an igraph object.}
\arguments{
\item{vertex_scale}{scaling function for the vertex sizes.}

\item{vertex_min}{minimum vertex size.}

\item{vertex_min}{maximum vertex size.}

\item{col_pal}{color palette to color the vertices by.}
}
\value{
an igraph object.
}
\description{
Exports the 1-skeleton to a graph using the igraph library.
}
\details{
By default, the vertices of the output graph are given "color", "size", and "label" 
attributes. The vertex colors are colored according on a blue to red rainbow according 
to their mean filter value (see \code{\link{bin_color}}). The vertex sizes are scaled according 
to the number of points they contain, scaled by \code{vertex_scale}, and bounded between 
(\code{vertex_min}, \code{vertex_max}). The vertex labels are in the format "<id>:<size>".\cr
\cr
The edges are colored similarly by the average filter value of the points intersecting
both nodes they connect too.
}
