% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ball_cover.R
\docType{class}
\name{BallCover}
\alias{BallCover}
\title{Ball Cover}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BallCover
}
\description{
This class provides a cover whose open sets are formed by the union of \deqn{\epsilon}-balls centered 
about each point. Using this class requires the \code{RANN} package to be installed, and thus explicitly assumes
the filter space endowed with the euclidean metric.
}
\section{Fields}{

\describe{
\item{\code{epsilon}}{:= radius of the ball to form around each point}
}}

\author{
Matt Piekenbrock
}
\keyword{datasets}
