% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallCover.R
\docType{class}
\name{BallCover}
\alias{BallCover}
\title{Ball Cover}
\description{
This class provides a cover whose open sets are formed by the union of \deqn{\epsilon}-balls centered 
about each point. Using this class requires the \code{RANN} package to be installed, and thus explicitly assumes
the filter space endowed with the euclidean metric.
}
\author{
Matt Piekenbrock
}
\section{Super class}{
\code{\link[Mapper:CoverRef]{Mapper::CoverRef}} -> \code{BallCover}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{epsilon}}{:= radius of the ball to form around each point}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BallCover-new}{\code{BallCover$new()}}
\item \href{#method-BallCover-validate}{\code{BallCover$validate()}}
\item \href{#method-BallCover-format}{\code{BallCover$format()}}
\item \href{#method-BallCover-construct_cover}{\code{BallCover$construct_cover()}}
\item \href{#method-BallCover-clone}{\code{BallCover$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Mapper" data-topic="CoverRef" data-id="neighborhood"><a href='../../Mapper/html/CoverRef.html#method-CoverRef-neighborhood'><code>Mapper::CoverRef$neighborhood()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BallCover-new"></a>}}
\if{latex}{\out{\hypertarget{method-BallCover-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BallCover$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BallCover-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BallCover-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BallCover$validate(filter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BallCover-format"></a>}}
\if{latex}{\out{\hypertarget{method-BallCover-format}{}}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BallCover$format(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BallCover-construct_cover"></a>}}
\if{latex}{\out{\hypertarget{method-BallCover-construct_cover}{}}}
\subsection{Method \code{construct_cover()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BallCover$construct_cover(filter, index = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BallCover-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BallCover-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BallCover$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
