% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo_booth.R
\name{check_in_interval}
\alias{check_in_interval}
\title{Check if a point is in an interval}
\usage{
check_in_interval(endpoints)
}
\arguments{
\item{endpoints}{A vector of the two endpoints of the interval. Must be in increasing order.}
}
\value{
A function that returns TRUE or FALSE depending on if an input data point is in between the interval's endpoints.
}
\description{
Check if a point is in an interval
}
