% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{compute_edges}
\alias{compute_edges}
\title{Computes the edges of the mapper.}
\arguments{
\item{which_level_pairs}{(n x 2) matrix of indices of the covers index set to check.}

\item{min_weight}{minimum intersection size to consider as an edge. Defaults to 1.}
}
\description{
Computes (or updates) the edges composing the topological graph (1-skeleton).
}
