% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_factory.R
\name{compute_tightness}
\alias{compute_tightness}
\title{Tightness Calculator}
\usage{
compute_tightness(dists, cluster)
}
\arguments{
\item{dists}{A distance matrix for points in the cluster.}

\item{cluster}{A list containing named vectors, whose names are data point names and whose values are cluster labels.}
}
\value{
A real number in \eqn{[0,1]} representing the mean distance to the medoid of the cluster, which is the data point with the smallest combined distance to every other point.
A smaller value indicates a tighter cluster based on this measure.
}
\description{
Compute a measure of dispersion for a single cluster.
}
