% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{construct_pullback}
\alias{construct_pullback}
\title{Constructs the (decomposed) pullback cover.}
\arguments{
\item{pullback_ids}{indices of the \code{\link[Mapper:CoverRef]{covers}} \code{index_set}, or \code{NULL}.}

\item{...}{additional parameters to pass to the \code{clustering_algorithm}.}
}
\description{
Executes the clustering algorithm on subsets of \code{X},
}
\details{
This methods uses the function given by \code{clustering_algorithm} field to 
decompose the preimages returned by the \code{cover} member into connected components, which are 
stored as \code{vertices}. Indices may be passed to limit which sets are decomposed, otherwise 
the sets in the cover all considered.
\cr
Note that this method removes the \code{vertices} associated with \code{pullback_ids}, but does not 
modify the simplicial complex.
}
\seealso{
\code{\link{construct_k_skeleton}} \code{\link{construct_nerve}} \code{\link{use_clustering_algorithm}}
}
