% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baskin_robbins.R
\name{convert_to_clusters}
\alias{convert_to_clusters}
\title{Convert Balls to Clusters}
\usage{
convert_to_clusters(bins)
}
\arguments{
\item{bins}{A \code{list} of bins, each containing a named vector of data points.}
}
\value{
A named vector whose names are data point names and whose values are cluster labels (\code{integer}s).
}
\description{
Perform trivial clustering on a set of balled data.
}
