% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterer_farm.R
\name{convert_to_clusters}
\alias{convert_to_clusters}
\title{"Clustering" for ballmapper just means treating each bin as its own cluster.}
\usage{
convert_to_clusters(bins)
}
\arguments{
\item{bins}{A list of bins, each containing names of data from some data frame.}
}
\value{
A named vector whose names are data point names and whose values are cluster labels
}
\description{
"Clustering" for ballmapper just means treating each bin as its own cluster.
}
