% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverRef.R
\name{CoverRef}
\alias{CoverRef}
\alias{cover}
\title{Cover abstract class}
\format{
An \code{\link{R6Class}} generator object
}
\description{
Reference Class (R6) implementation of a Cover. This class is meant to act as an abstract class to derive other 
types of covering generators with. Minimally, a derived covering class must implement the
'construct_cover' method to populate the 'level_sets' list with point indices, and any parameters 
that the derived class requires. 

Additional methods may also be added to improve the efficiency of the cover. 
See the \href{https://peekxc.github.io/Mapper/articles/UsingCustomCover.html}{vignette} on creating a custom 
cover for details.
}
\section{Fields}{

The following is a list of the fields available for derived classes. Each may be accessed 
by the \code{self} environment. See \code{?R6} for more details. 
\itemize{
 \item{\emph{level_sets}:}{ named list, indexed by \code{index_set}, whose values represent indices in the original data set to cluster over.}
 \item{\emph{index_set}:}{ character vector of keys that uniquely index the open sets of the cover.}
 \item{\emph{typename}:}{ unique string identifier identifying the covering method.}
}
}

\author{
Matt Piekenbrock
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoverRef-new}{\code{CoverRef$new()}}
\item \href{#method-CoverRef-format}{\code{CoverRef$format()}}
\item \href{#method-CoverRef-construct_cover}{\code{CoverRef$construct_cover()}}
\item \href{#method-CoverRef-neighborhood}{\code{CoverRef$neighborhood()}}
\item \href{#method-CoverRef-validate}{\code{CoverRef$validate()}}
\item \href{#method-CoverRef-clone}{\code{CoverRef$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoverRef-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoverRef-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverRef$new(typename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoverRef-format"></a>}}
\if{latex}{\out{\hypertarget{method-CoverRef-format}{}}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverRef$format(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoverRef-construct_cover"></a>}}
\if{latex}{\out{\hypertarget{method-CoverRef-construct_cover}{}}}
\subsection{Method \code{construct_cover()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverRef$construct_cover(index = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoverRef-neighborhood"></a>}}
\if{latex}{\out{\hypertarget{method-CoverRef-neighborhood}{}}}
\subsection{Method \code{neighborhood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverRef$neighborhood(filter, k = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoverRef-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CoverRef-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverRef$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoverRef-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoverRef-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverRef$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
