% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverRef.R
\docType{class}
\name{CoverRef}
\alias{CoverRef}
\title{Cover abstract class}
\format{An \code{\link{R6Class}} generator object}
\usage{
CoverRef
}
\description{
Reference Class (R6) implementation of a Cover. This class is meant to act as an abstract class to derive other 
types of covering generators with. Minimally, a derived covering class must implement the
'construct_cover' method to populate the 'level_sets' list with point indices, and any parameters 
that the derived class requires. 

Additional methods may also be added to improve the efficiency of the cover. See the vignette on creating a custom 
cover for details.
}
\section{Fields}{

\describe{
\item{\code{typename}}{Unique string identifier for the covering.}

\item{\code{index_set}}{character vector used to index the 'level_sets' list}

\item{\code{level_sets}}{list of the point in the preimages of the sets comprising the cover, indexed by the index set}
}}

\section{Private variables}{

The following is a list of private variables available for derived classes. Each may be accessed 
by the \code{private} environment. See \code{?R6} for more details. 
\itemize{
 \item{\emph{.level_sets}}{ named list, indexed by \code{.index_set}, whose values represent indices in the original data set to cluster over.}
 \item{\emph{.index_set}}{ character vector of keys that uniquely index the level sets.}
 \item{\emph{.typename}}{ unique string identifier of the covering method.}
}
}

\author{
Matt Piekenbrock
}
\keyword{datasets}
