% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverRef.R
\docType{data}
\name{CoverRef}
\alias{CoverRef}
\alias{cover}
\title{Cover abstract class}
\format{An \code{\link{R6Class}} generator object}
\usage{
CoverRef
}
\description{
Reference Class (R6) implementation of a Cover. This class is meant to act as an abstract class to derive other 
types of covering generators with. Minimally, a derived covering class must implement the
'construct_cover' method to populate the 'level_sets' list with point indices, and any parameters 
that the derived class requires. 

Additional methods may also be added to improve the efficiency of the cover. 
See the \href{https://peekxc.github.io/Mapper/articles/UsingCustomCover.html}{vignette} on creating a custom 
cover for details.
}
\section{Fields}{

The following is a list of the fields available for derived classes. Each may be accessed 
by the \code{self} environment. See \code{?R6} for more details. 
\itemize{
 \item{\emph{typename}:}{ name of the covering method.}
 \item{\emph{sets}:}{ named list, indexed by \code{index_set}, representing the collection of sets.}
 \item{\emph{index_set}:}{ character vector of keys that uniquely index the \code{sets} of the cover.}
}
}

\author{
Matt Piekenbrock
}
\keyword{datasets}
