% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_ref.R
\name{covers_available}
\alias{covers_available}
\title{Print available covers}
\usage{
covers_available()
}
\description{
Prints the covering generators available with with package, along with their parameters.
}
