% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{create_1D_mapper_object}
\alias{create_1D_mapper_object}
\title{Run 1D mapper}
\usage{
create_1D_mapper_object(
  data,
  dists,
  filtered_data,
  cover,
  clustering_method = "single"
)
}
\arguments{
\item{data}{A data frame.}

\item{dists}{A distance matrix for the data frame.}

\item{filtered_data}{The result of a function applied to the data frame; there should be one row per observation in the original data frame.}

\item{cover}{A 2D array of interval left and right endpoints.}

\item{clustering_method}{Your favorite clustering algorithm.}
}
\value{
A list of two dataframes, one with node data containing bin membership,
datapoints per cluster, and cluster dispersion, and one with edge data
containing sources, targets, and weights representing overlap strength.
}
\description{
Run mapper using a one-dimensional filter, a cover of intervals, and a clustering algorithm.
}
