% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baskin_robbins.R
\name{create_ball_mapper_object}
\alias{create_ball_mapper_object}
\title{Run Mapper using a trivial filter, a cover of balls, and no clustering algorithm.}
\usage{
create_ball_mapper_object(data, dists, eps)
}
\arguments{
\item{data}{A data frame.}

\item{dists}{A distance matrix for the data frame.}

\item{eps}{A positive real number for your desired ball radius.}
}
\value{
A list of two data frames, \code{nodes} and \code{edges}, which contain information about the Mapper graph constructed from the given parameters.

The node data frame consists of:
\itemize{
\item \code{id}: vertex ID
\item \code{cluster_size}: number of data points in vertex
\item \code{mean_dist_to_medoid}: mean distance to medoid of vertex
\item \code{data}: names of data points in cluster
}

The \code{edge} data frame contains consists of:
\itemize{
\item \code{source}: vertex ID of edge source
\item \code{target}: vertex ID of edge target
\item \code{weight}: Jaccard index of edge; this is the size of the intersection between the vertices divided by the union
\item \code{overlap_data}: names of data points in overlap
\item \code{overlap_size}: number of data points overlap
}
}
\description{
Run Mapper using an \eqn{\varepsilon}-net cover (greedily generated) and the 2D inclusion function as a filter.
}
\examples{
data = data.frame(x = sapply(1:100, function(x) cos(x)), y = sapply(1:100, function(x) sin(x)))
eps = .5

create_ball_mapper_object(data, dist(data), eps)
}
