% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{create_bins}
\alias{create_bins}
\title{Create bins of data}
\usage{
create_bins(data, filtered_data, cover_element_tests)
}
\arguments{
\item{data}{A data frame.}

\item{filtered_data}{The result of a function applied to the data frame; there should be one row per observation in the original data frame.}

\item{cover_element_tests}{A list of membership test functions for a list of cover elements. Each member of \code{cover_element_tests} should be able to identify (return \code{TRUE} or \code{FALSE}) if a single input data point is a member of the cover element it represents.}
}
\value{
A list of bins, each containing a vector of the names of the data inside it.
}
\description{
Create bins of data
}
