% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baskin_robbins.R
\name{create_clusterball_mapper_object}
\alias{create_clusterball_mapper_object}
\title{Run clusterball Mapper}
\usage{
create_clusterball_mapper_object(
  data,
  dist1,
  dist2,
  eps,
  clusterer = local_hierarchical_clusterer("single")
)
}
\arguments{
\item{data}{A data frame.}

\item{dist1}{A distance matrix for the data frame; this will be used to ball the data.}

\item{dist2}{Another distance matrix for the data frame; this will be used to cluster the data after balling.}

\item{eps}{A positive real number for your desired ball radius.}

\item{clusterer}{A function which accepts a list of distance matrices as input, and returns the results of clustering done on each distance matrix.}
}
\value{
A list of two data frames, \code{nodes} and \code{edges}, which contain information about the Mapper graph constructed from the given parameters.

The node data frame consists of:
\itemize{
\item \code{id}: vertex ID
\item \code{cluster_size}: number of data points in vertex
\item \code{mean_dist_to_medoid}: mean distance to medoid of vertex
\item \code{data}: names of data points in cluster
\item \code{patch}: level set ID
}

The \code{edge} data frame contains consists of:
\itemize{
\item \code{source}: vertex ID of edge source
\item \code{target}: vertex ID of edge target
\item \code{weight}: Jaccard index of edge; this is the size of the intersection between the vertices divided by the union
\item \code{overlap_data}: names of data points in overlap
\item \code{overlap_size}: number of data points overlap
}
}
\description{
Run ball Mapper, but additionally cluster within the balls. Can use two different distance matrices to accomplish this.
}
\examples{
data = data.frame(x = sapply(1:100, function(x) cos(x)), y = sapply(1:100, function(x) sin(x)))
data.dists = dist(data)
eps = 1

create_clusterball_mapper_object(data, data.dists, data.dists, eps)
}
