% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{create_mapper_object}
\alias{create_mapper_object}
\title{Create a mapper object}
\usage{
create_mapper_object(
  data,
  dists,
  filtered_data,
  cover_element_tests,
  method = "none"
)
}
\arguments{
\item{data}{A data frame.}

\item{dists}{A distance matrix for the data frame.}

\item{filtered_data}{The result of a function applied to the data frame; there should be one row per observation in the original data frame.}

\item{cover_element_tests}{A list of membership test functions for a list of cover elements. Each member of \code{cover_element_tests} should be able to identify (return \code{TRUE} or \code{FALSE}) if a single input data point is a member of the cover element it represents.}

\item{method}{The desired clustering method to use. e.g., "single"}
}
\value{
A list of two dataframes, one with node data containing bin membership,
datapoints per cluster, and cluster dispersion, and one with edge data
containing sources, targets, and weights representing overlap strength.
}
\description{
Run the mapper algorithm on a data frame.
}
