% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{create_mapper_object}
\alias{create_mapper_object}
\title{Create a mapper object}
\usage{
create_mapper_object(
  data,
  dists,
  filtered_data,
  cover_element_tests,
  method = "none"
)
}
\arguments{
\item{data}{A data frame.}

\item{dists}{A distance matrix for the data frame.}

\item{filtered_data}{The result of a function applied to the data frame; there should be one row per observation in the original data frame.}

\item{cover_element_tests}{A list of membership test functions for a list of cover elements. Each member of \code{cover_element_tests} should be able to identify (return \code{TRUE} or \code{FALSE}) if a single input data point is a member of the cover element it represents.}

\item{method}{The desired clustering method to use. e.g., "single"}
}
\value{
A list of two dataframes, one with node data containing bin membership,
datapoints per cluster, and cluster dispersion, and one with edge data
containing sources, targets, and weights representing overlap strength.
}
\description{
Run the mapper algorithm on a data frame.
}
\examples{
num_points = 5000

# generate some points close to a curve in 3D space
P.data = data.frame(
  x = sapply(1:num_points, function(x)
    sin(x) * 10) + rnorm(num_points, 0, 0.1),
  y = sapply(1:num_points, function(x)
    cos(x) ^ 2 * sin(x) * 10) + rnorm(num_points, 0, 0.1),
  z = sapply(1:num_points, function(x)
    10 * sin(x) ^ 2 * cos(x)) + rnorm(num_points, 0, 0.1)
)

# create distance matrix
P.dist = dist(P.data)

# lens function is projection to the \eqn{x}-coordinate
projx = P.data$x

# cover parameters to generate a width-balanced cover
num_bins = 10
percent_overlap = 25

# generate the cover
xcover = create_width_balanced_cover(min(projx), max(projx), num_bins, percent_overlap)

# each of the "cover" elements will really be a function that checks if a data point lives in it
xcovercheck = apply(xcover, 1, check_in_interval)

# build the mapper object
xmapper = create_mapper_object(
  data = P.data,
  dists = P.dist,
  filtered_data = projx,
  cover_element_tests = xcovercheck,
  method = "single"
)
}
