% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arboretum.R
\name{cut_dendrogram}
\alias{cut_dendrogram}
\title{Cut a dendrogram}
\usage{
cut_dendrogram(dend, threshold)
}
\arguments{
\item{dend}{A single dendrogram.}

\item{threshold}{A mininum tallest branch value.}
}
\value{
A named vector whose names are data point names and whose values are cluster labels. The number of clusters is determined to be 1 if the tallest branch of the dendrogram is less than the threshold, or if the index of dispersion (standard deviation squared divided by mean) of the branch heights is too low. Otherwise, we cut at the longest branch of the dendrogram to determine the number of clusters.
}
\description{
Cut a dendrogram
}
