% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterer_farm.R
\name{cut_dendrogram}
\alias{cut_dendrogram}
\title{Cut a single dendrogram}
\usage{
cut_dendrogram(dend, cut_height)
}
\arguments{
\item{dend}{A dendrogram to be cut.}

\item{cut_height}{A height at which to cut the dendrogram.}
}
\value{
A cluster assignment resulting from \code{cutree}. It is a named vector whose names are data point names and whose values are cluster labels.
}
\description{
Cut a single dendrogram
}
