% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_methods.R
\name{cut_dendrogram}
\alias{cut_dendrogram}
\title{Cut a dendrogram in context}
\usage{
cut_dendrogram(dend, threshold)
}
\arguments{
\item{dend}{A single dendrogram.}

\item{threshold}{A mininum tallest branch value.}
}
\value{
A named vector whose names are data point names and whose values are cluster labels.
}
\description{
Cut a dendrogram in context
}
\details{
The number of clusters is determined to be 1 if the tallest branch of the dendrogram is less than the threshold, or if the index of dispersion (standard deviation squared divided by mean) of the branch heights is below 0.015. Otherwise, we cut at the longest branch of the dendrogram to determine the number of clusters.
}
