% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_methods.R
\name{cutoff_first_bin}
\alias{cutoff_first_bin}
\title{cutoff_first_bin function}
\usage{
cutoff_first_bin(hcl, num_bins)
}
\arguments{
\item{hcl}{hierarchical clustering in the form an 'hclust' object.}

\item{num_bins}{Controls how many bins there are in the histogram used to determine cutoff.}
}
\description{
Uses the heuristic from the original Mapper paper to cut a given hierarchical cluster tree at the
threshold corresponding to the first empty of a histogram of the linkage distances.
}
