% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballmapper.R
\name{cyballmapper}
\alias{cyballmapper}
\title{Runs ballmapper and passes data to Cytoscape for visualization.}
\usage{
cyballmapper(data, dists, eps)
}
\arguments{
\item{data}{Your input data. Ideally a dataframe.}

\item{dists}{A distance matrix for your data. Can be a `dist` object or 2D matrix.}

\item{eps}{A positive real number for your desired ball radius.}
}
\description{
Runs ballmapper and passes data to Cytoscape for visualization.
}
\examples{
circle.data = data.frame( x= sapply(1:1000, function(x) cos(x)) + rnorm(100, 500, .03), y = sapply(1:1000, function(x) sin(x)) + rnorm(100, 0, 0.03))
circle.dist = dist(circle.data)

# make sure Cytoscape is running in the background or this will not work
cyballmapper(circle.data, circle.dist, .3)
}
