% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{cymapper}
\alias{cymapper}
\title{Runs 1D mapper and passes data to Cytoscape for visualization.}
\usage{
cymapper(mapperobject)
}
\arguments{
\item{data}{A dataframe.}

\item{filtered_data}{A single column of the input data.}

\item{dists}{A distance matrix containing pairwise relations among the input data. Can be a \code{dist} object or 2D matrix.}

\item{num_bins}{The number of "bins" to split the input data into based on the filter. A positive integer.}

\item{percent_overlap}{The percent overlap desired between each "bin." An integer between 0 and 100 (inclusive).}

\item{clustering_method}{Desired clustering method. A string from these options: "single" (single-linkage hierarchical)}
}
\description{
Runs 1D mapper and passes data to Cytoscape for visualization.
}
