% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1D_mapper.R
\name{cymapper}
\alias{cymapper}
\title{Runs 1D mapper and passes data to Cytoscape for visualization.}
\usage{
cymapper(
  data,
  filtered_data,
  dists,
  num_bins,
  percent_overlap,
  clustering_method
)
}
\arguments{
\item{data}{Your input data. Ideally a dataframe.}

\item{filtered_data}{A single column of your data, of numeric value.}

\item{dists}{A distance matrix for your data. Can be a `dist` object or 2D matrix.}

\item{num_bins}{A positive integer for the number of bins you want to split up your filtered data into.}

\item{percent_overlap}{The percent (0-100) overlap you want between each bin and its neighbors.}

\item{clustering_method}{Desired clustering method. A string from these options: "single" (single-linkage)}
}
\description{
Runs 1D mapper and passes data to Cytoscape for visualization.
}
\examples{
circle.data = data.frame( x= sapply(1:1000, function(x) cos(x)) + rnorm(100, 500, .03), y = sapply(1:1000, function(x) sin(x)) + rnorm(100, 0, 0.03))
circle.dist = dist(circle.data)

# make sure Cytoscape is running in the background or this will not work
cymapper(circle.data, circle.data$x, circle.dist, 10, 15, "single")
}
