% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{dashboard}
\alias{dashboard}
\title{Mapper dashboard}
\usage{
dashboard(mapper_obj, X = mapper_obj$X, node_color_f = "default")
}
\arguments{
\item{mapper_obj}{MapperRef object.}

\item{X}{data.frame to display in lower table and subset on node selection.}

\item{node_color_f}{List of color functions. See details.}
}
\description{
Starts an interactive dashboard for analyzing Mapper constructions using Shiny.
}
\details{
The data.frame supplied by the \code{X} argument is displayed in an interactive table in the dashboard.
By default, if \code{X} is not supplied it is assigned the original data the Mapper object was constructed with.
This can be helpful in situations where the point cloud data Mapper was constructed required preprocessing (such as scaling), whereas the original,
unscaled data.frame is better suited for direct analysis.

If supplied, \code{node_color_f} must be a named list of functions which accept as input a MapperRef object and return as output a vector of numerical values, one per node.
These values are automatically converted to color hex codes based on the current palette. If this isn't supplied, a few default color functions are supplied.
}
