% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{exportMapper}
\alias{exportMapper}
\title{Exports minimal information about the Mapper}
\arguments{
\item{graph_type}{export preference on the structure the graph output.}
}
\value{
List with the following members: 
\itemize{
  \item \emph{vertices} list of the indices of the original data the current vertex intersects with.
  \item \emph{graph} some adjacency representation of the mapper graph.
  \item \emph{pullback} map connecting the index set of the cover and the vertices.
}
}
\description{
This function exports a few core characteristics about the mapper complex as a list.
}
\details{
\code{graph_type} must be one of 'adjacency_matrix', 'adjacency_list', or 'edgelist'.
}
