% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{exportMapper}
\alias{exportMapper}
\title{Exports mapper information}
\arguments{
\item{graph_type}{export preference on the structure the graph output.}
}
\value{
list with the following members: 
\itemize{
\item \emph{vertices} list of the indices of the original data the current vertex intersects with.
\item \emph{graph} some adjacency representation of the mapper graph.
\item \emph{level_sets} map connecting the which vertices belong to the preimage of the sets in the cover. 
}
}
\description{
Exports mapper information
}
