% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedIntervalCover.R
\docType{class}
\name{FixedIntervalCover}
\alias{FixedIntervalCover}
\title{Fixed Interval Cover}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FixedIntervalCover
}
\description{
\code{\link{R6}} class representing a fixed interval cover.
}
\details{
The fixed interval cover is multidimensional, two-parameter family of covers. Given the number of
intervals and the overlap percentage between these intervals, this class constructs cover of the
given filter space, where the open sets represent a (generalized) collection of interval level sets, 
distributed uniformly in a grid-like fashion. Unlike the \code{\link{RestrainedIntervalCover}}, the open sets
are not constrained by range of the filter values. \cr 
\cr 
The mapper corresponding to this cover may be thought of as a relaxed Reeb graph.
}
\section{Fields}{

\describe{
\item{\code{number_intervals}}{vector of number of bins to cover the Z with (per dimension)}

\item{\code{percent_overlap}}{vector of overlap percentages}
}}

\seealso{
Other cover: \code{\link{ArbitraryCover}},
  \code{\link{BallCover}}, \code{\link{IndexedCover}}
}
\author{
Matt Piekenbrock
}
\concept{cover}
\keyword{datasets}
