% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_imapper.R
\name{future_imapper}
\alias{future_imapper}
\title{imapping in Parallel and Ending parallel imapping \code{furrr} in R}
\usage{
future_imapper(...)
}
\arguments{
\item{workers}{default parallel workers are detectCores()-2}
}
\description{
imapping in Parallel and Ending parallel imapping \code{furrr} in R
}
