% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_invoke_mapper.R
\name{future_invoke_mapper}
\alias{future_invoke_mapper}
\title{MAPping in Parallel and Ending parallel mapping \code{furrr} in R}
\arguments{
\item{workers}{default parallel workers are detectCores()-2}
}
\description{
MAPping in Parallel and Ending parallel mapping \code{furrr} in R
}
