% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_invoke_mapper.R
\docType{data}
\name{future_invoke_mapper}
\alias{future_invoke_mapper}
\title{future_invoke_mapper}
\format{
An object of class \code{environment} of length 39.
}
\usage{
future_invoke_mapper
}
\description{
future_invoke_mapper
}
\keyword{datasets}
