% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureM.R
\name{future_mapper}
\alias{future_mapper}
\alias{future_mapper1}
\title{future_mapper}
\usage{
future_mapper1(...)
}
\arguments{
\item{...}{Input parameter}
}
\value{
Output result
}
\description{
future_mapper
}
