% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugeProp.R
\name{gaugeProp}
\alias{gaugeProp}
\title{Gauge proportion}
\usage{
gaugeProp(intersectPoly)
}
\arguments{
\item{intersectPoly}{Shapefile of the intersected catchment polygon}

\item{catchment}{Catchment shapefile}
}
\value{
Rainfall proportions for each rain gauge
}
\description{
Calculates the the proportion of catchment that the gauge occupies using Thiessen / Voronoi polygons
}
\examples{
library(riskyData)
data(crowle); data(bickley); data(barnhurst); data(hollies); data(ledbury);
data(bettwsYCrwyn); data(bewdCatch)

# Obtain gauge coordinates
gcs <- getCoords(crowle,
                 bickley,
                 barnhurst,
                 hollies,
                 ledbury,
                 bettwsYCrwyn)

bewdTeeSun <- teeSun(gaugeCoords = gcs, catchment = bewdCatch)
int <- intersectPoly(coords = gcs,
                     voronoi = bewdTeeSun,
                     catchment = bewdCatch)
gaugeProp(intersectPoly = int)
}
