% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{get_1D_mapper_data}
\alias{get_1D_mapper_data}
\title{Run 1D mapper}
\usage{
get_1D_mapper_data(
  data,
  filtered_data,
  dists,
  num_bins,
  percent_overlap,
  clustering_method
)
}
\arguments{
\item{data}{A dataframe.}

\item{filtered_data}{A single column of the input data.}

\item{dists}{A distance matrix containing pairwise relations among the input data. Can be a \code{dist} object or 2D matrix.}

\item{num_bins}{The number of "bins" to split the input data into based on the filter. A positive integer.}

\item{percent_overlap}{The percent overlap desired between each "bin." An integer between 0 and 100 (inclusive).}

\item{clustering_method}{Desired clustering method. A string from these options: "single" (single-linkage hierarchical)}
}
\value{
A list containing:
\itemize{
\item nodes - A dataframe containing vertex/cluster information:
\itemize{
\item id - The vertex label.
\item size - How many datapoints are contained in that vertex.
\item tightness - A measure of dispersion of the data inside the vertex.
\item data - A string containing names of the datapoints within the vertex.
\item bin - Which bin the vertex belongs to.
}
\item edges - A dataframe containing vertex/cluster overlap information:
\itemize{
\item source - One endpoint of an edge.
\item target - The other endpoint of an edge.
\item weight - A measure of the significance of the edge.
}
}
}
\description{
Run mapper using a 1-dimensional projection filter and a width-balanced cover.
}
