% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{get_ballmapper_data}
\alias{get_ballmapper_data}
\title{Run ballmapper}
\usage{
get_ballmapper_data(data, dists, eps)
}
\arguments{
\item{data}{Your input data. Ideally a dataframe.}

\item{dists}{A distance matrix for your data. Can be a \code{dist} object or 2D matrix.}

\item{eps}{A positive real number for your desired ball radius.}
}
\value{
A list of two dataframes, one with node data containing ball size,
datapoints per ball, ball tightness, and one with edge data
containing sources, targets, and weights representing overlap strength.
}
\description{
Run mapper using an \eqn{\varepsilon}-net cover (greedily generated) and the 2D inclusion function as a filter.
}
