% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watering_hole.R
\name{get_cluster_tightness_vector}
\alias{get_cluster_tightness_vector}
\title{Compute dispersion measures of a list of clusters}
\usage{
get_cluster_tightness_vector(dists, binclust_data)
}
\arguments{
\item{dists}{A distance matrix for the data points inside the input clusters}

\item{binclust_data}{A list of bins, each containing named vectors whose names are those of data points and whose values are cluster ids}
}
\value{
A vector of real numbers in \eqn{(0,\infty)} representing a measure of dispersion of a cluster, calculated according to \code{\link[=compute_tightness]{compute_tightness()}}
}
\description{
Compute dispersion measures of a list of clusters
}
