% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_factory.R
\name{get_cluster_tightness_vector}
\alias{get_cluster_tightness_vector}
\title{Tightnesses Calculator}
\usage{
get_cluster_tightness_vector(dists, binclust_data)
}
\arguments{
\item{dists}{A distance matrix for the data points inside all the input clusters.}

\item{binclust_data}{A list of named vectors whose names are those of data points and whose values are cluster IDs (integers).}
}
\value{
A vector of real numbers in \eqn{(0,\infty)} containing mean distances to the medoids of each cluster in \code{dists}.
}
\description{
Tightnesses Calculator
}
