% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_factory.R
\name{get_clusters}
\alias{get_clusters}
\title{Varnished Clustering}
\usage{
get_clusters(bins, dists, clusterer)
}
\arguments{
\item{bins}{A \code{list} containing "bins" of vectors of names of data points.}

\item{dists}{A distance matrix containing pairwise distances between named data points. Needs to be acceptable to \code{clusterer}.}

\item{clusterer}{A function which accepts a list of distance matrices as input, and returns the results of clustering done on each distance matrix.}
}
\value{
The output of \code{clusterer} applied to a list of distance matrices, which should be a list containing named vectors (one per bin), whose names are data point names and whose values are cluster labels.
These labels are unique to the bin which has them.
}
\description{
Process level sets of data and a global distance matrix to return fresh clusters.
}
