% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{get_edgelist_from_overlaps}
\alias{get_edgelist_from_overlaps}
\title{Edge List Maker}
\usage{
get_edgelist_from_overlaps(overlaps, num_vertices)
}
\arguments{
\item{overlaps}{A named \code{list} of edges, whose elements contain the names of clusters in the overlap represented by that edge; i.e., the output of \code{\link[=get_overlaps]{get_overlaps()}}.
The names should be integers corresponding to their ordering among all possible pairs of vertices.}

\item{num_vertices}{The number of vertices in the graph.}
}
\value{
A 2D array representing the edge list of a graph.
}
\description{
Turn the edge labels given by \code{\link[=get_overlaps]{get_overlaps()}} into a usable edge list.
}
