% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterer_farm.R
\name{get_hierarchical_clusters}
\alias{get_hierarchical_clusters}
\title{Do Much Clustering}
\usage{
get_hierarchical_clusters(dist_mats, method, cut_height = -1)
}
\arguments{
\item{dist_mats}{A list of distance matrices to be used for clustering. Must be a \code{list} of dist` objects.}

\item{method}{A string to pass to \link{hclust} to tell it what kind of clustering to do. Can be \code{single}, \code{complete}, etc.}

\item{cut_height}{A global cut height. If not specified or negative, dendrograms will be cut individually.}
}
\value{
A \code{list} containing named vectors (one per dendrogram), whose names are data point names and whose values are cluster labels (integers).
}
\description{
Perform hierarchical clustering on distance matrices.
}
