% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_factory.R
\name{get_hierarchical_clusters}
\alias{get_hierarchical_clusters}
\title{Perform hierarchical clustering and process dendrograms}
\usage{
get_hierarchical_clusters(dist_mats)
}
\arguments{
\item{dist_mats}{A list of distance matrices to be used for clustering.}
}
\value{
A list containing named vectors (one per dendrogram), whose names are data point names and whose values are cluster labels.
}
\description{
Perform hierarchical clustering and process dendrograms
}
