% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arboretum.R
\name{get_tallest_branch}
\alias{get_tallest_branch}
\title{Find the tallest branch of a dendrogram}
\usage{
get_tallest_branch(dend)
}
\arguments{
\item{dend}{A single dendrogram.}
}
\value{
The height of the tallest branch (longest time between merge heights) of the input dendrogram.
}
\description{
Find the tallest branch of a dendrogram
}
