% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gromov_hausdorff.R
\name{gromov_hausdorff}
\alias{gromov_hausdorff}
\title{gromov hausdorff}
\usage{
gromov_hausdorff(d_X, d_Y, mu_X, mu_Y, flb_solver = "any",
  flb_only = FALSE, roi_opt = "default", control = "default",
  return_optimizer = FALSE)
}
\arguments{
\item{d_X}{either a dist object, or (n x n) numeric metric distance matrix over \code{X}.}

\item{d_Y}{either a dist object, or (m x m) numeric metric distance matrix over \code{Y}.}

\item{mu_X}{(n-length) numeric vector giving the probability measure at each point of \code{X}.}

\item{mu_Y}{(m-length) numeric vector giving the probability measure at each point of \code{Y}.}

\item{flb_solver}{Which LOP ROI plugin solver to use to solve the lower bound approximation.}

\item{flb_only}{If only crude approximation is needed, the lower bound can be returned. See details.}

\item{roi_opt}{Outer options to pass to ROI_solve}

\item{control}{Inner options to pass to ROI_solves 'control' parameter.}
}
\description{
This function implements the discrete, extended version of gromov hausdorff distance proposed by 
Memoli in section 7 of (1) for comparing two measure-metric spaces.
}
\details{
This function provides an implementation of the computational objective \emph{(Pp)} listed in Section 7 of (1), which 
calculates an Lp approximition of Gromov-Hausdorff distance between two measure-metric spaces. The first lower bound (FLB)
is first computed via an LOP, and then the solution is handed off as the starting point to the corresponding nonlinear optimization 
routine. This code utilizes the \emph{R Optimization Infrastructure}(\pkg{ROI}), and relies on having available at least one plugin 
available for to solve the initial lower bound (a linear program) and the final optimization (cast as a generic nonlinear optimization). 
See examples.
}
\references{
1. Mémoli, Facundo. "On the use of Gromov-Hausdorff distances for shape comparison." (2007).
}
