% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_imapper.R
\name{imapper_dbl}
\alias{imapper_dbl}
\title{imapper_dbl}
\usage{
imapper_dbl(..., workers = NULL)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
imapper_dbl
}
