% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_imapper.R
\name{imapper_lgl}
\alias{imapper_lgl}
\title{imapper_lgl}
\usage{
imapper_lgl(workers = parallel::detectCores() - 2, ...)
}
\arguments{
\item{workers}{default value is the \verb{total cores - 2}}
}
\description{
imapper_lgl
}
