% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectPoly.R
\name{intersectPoly}
\alias{intersectPoly}
\title{Intersect of polygons}
\usage{
intersectPoly(voronoi, catchment, coords)
}
\arguments{
\item{voronoi}{Voronoi/Thiessen polygon}

\item{catchment}{Catchment shapefile}

\item{coords}{Coordinates of rain gauges set as a polygon shapefile}
}
\value{
The calculated intersections of a polygon
}
\description{
Creates intersected polygons of catchment against Voronoi polygons
}
\examples{
library(riskyData)
data(crowle); data(bickley); data(barnhurst); data(hollies); data(ledbury);
data(bettwsYCrwyn); data(bewdCatch)

# Obtain gauge coordinates
gcs <- getCoords(crowle,
                 bickley,
                 barnhurst,
                 hollies,
                 ledbury,
                 bettwsYCrwyn)

bewdTeeSun <- teeSun(gaugeCoords = gcs, catchment = bewdCatch)
int <- intersectPoly(coords = gcs,
                     voronoi = bewdTeeSun,
                     catchment = bewdCatch)
}
