% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalDualTessellationCover.R
\docType{class}
\name{IntervalDualTessellationCover}
\alias{IntervalDualTessellationCover}
\title{Interval Dual Tessellation Cover}
\description{
[R6] class representing an interval dual tessellation cover.
}
\details{
The regular interval tessellation \eqn{T} of \eqn{Z} (the real line, or a
segment of it) consists of intervals of fixed length, shared endpoints, and
empty overlap. In order to completely partition \eqn{Z}, each interval is
considered closed at one end and open at the other. The dual tessellation
\eqn{T'} of \eqn{T} is obtained by taking the endpoints of the intervals at
the centers of the intervals of \eqn{T}. The union \eqn{C} of \eqn{T} and
\eqn{T'} is an _interval dual tessellation cover_.

An interval dual tessellation cover \eqn{C} is a special case of a
`[FixedIntervalCover]`, for a 1-dimensional lens and with `percent_overlap =
50`. \eqn{C} is _gomic_ (a **g**eneric, **o**pen, **m**inimal, **i**nterval
**c**over) in the sense of Carriere and Oudot (2016). \eqn{C} has constant
"depth", in the sense that every point in \eqn{Z} is a member of exactly two
sets in \eqn{C}. As discussed by Carriere and Oudot, because the mapper
construction collapses sets in the pullback cover to vertices of the nerve
complex, only 1-dimensional features of a point cloud \eqn{X} can be
recovered from a 1-dimensional lens \eqn{f:X->Z}. A cover of depth 2 is
sufficient to recover these features, so no tessellation covers of greater
depth are available. Every non-empty intersection of cover sets has the same
length.

`IntervalDualTessellationCover` takes the center of \eqn{f(X)} to be the
center of the overlap between a pair of intervals in \eqn{T} and in \eqn{T'}.
Thus, the tessellations have negation symmetry and the same number of sets.
When `width` is sufficiently large (twice the range of the data), the total
number of intervals is minimized at 2.
}
\seealso{
Other cover: 
\code{\link{FixedIntervalCover}},
\code{\link{SquareDualTessellationCover}},
\code{\link{TriangleDualTessellationCover}}
}
\author{
Jason Cory Brunson
}
\concept{cover}
\section{Super class}{
\code{\link[Mapper:CoverRef]{Mapper::CoverRef}} -> \code{IntervalDualTessellationCover}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{width}}{A positive number. The width of each interval in the cover, as a
fraction of the range of the lensed data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IntervalDualTessellationCover$new()}}
\item \href{#method-validate}{\code{IntervalDualTessellationCover$validate()}}
\item \href{#method-format}{\code{IntervalDualTessellationCover$format()}}
\item \href{#method-interval_bounds}{\code{IntervalDualTessellationCover$interval_bounds()}}
\item \href{#method-construct_index_set}{\code{IntervalDualTessellationCover$construct_index_set()}}
\item \href{#method-construct_cover}{\code{IntervalDualTessellationCover$construct_cover()}}
\item \href{#method-neighborhood}{\code{IntervalDualTessellationCover$neighborhood()}}
\item \href{#method-clone}{\code{IntervalDualTessellationCover$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalDualTessellationCover$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalDualTessellationCover$validate(filter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalDualTessellationCover$format(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-interval_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-interval_bounds}{}}}
\subsection{Method \code{interval_bounds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalDualTessellationCover$interval_bounds(filter, index = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-construct_index_set"></a>}}
\if{latex}{\out{\hypertarget{method-construct_index_set}{}}}
\subsection{Method \code{construct_index_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalDualTessellationCover$construct_index_set(filter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-construct_cover"></a>}}
\if{latex}{\out{\hypertarget{method-construct_cover}{}}}
\subsection{Method \code{construct_cover()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalDualTessellationCover$construct_cover(filter, index = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-neighborhood"></a>}}
\if{latex}{\out{\hypertarget{method-neighborhood}{}}}
\subsection{Method \code{neighborhood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalDualTessellationCover$neighborhood(filter, k)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalDualTessellationCover$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
