\name{mappeR-package}
\alias{mappeR-package}
\docType{package}
\title{
mappeR
}
\description{
Implements the Mapper algorithm, for low-dimensional representations of topological structure in (possibly) high-dimensional data.
}
\details{
\tabular{ll}{
Package: \tab mappeR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-01-31\cr
License: \tab GPL-3\cr
}
}
\author{
Maintainer: Matthew Arnold <matthew.arnold@bristol.ac.uk>
}
\references{
Singh, G., Memoli, F., Carlsson, G. (2007) Topological Methods for the Analysis of High Dimensional Data Sets and 3D Object Recognition
}
\keyword{ package }
\examples{
}