% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappeR-package.R
\docType{package}
\name{mappeR-package}
\alias{mappeR}
\alias{mappeR-package}
\title{mappeR: Construct and Visualize TDA Mapper Graphs}
\description{
Topological data analysis (TDA) is a method of data analysis that uses techniques from topology to analyze high-dimensional data. Here we implement Mapper, an algorithm from this area developed by Singh, Mémoli and Carlsson (2007) which generalizes the concept of a Reeb graph \url{https://en.wikipedia.org/wiki/Reeb_graph}. The output graph is able to be visualized in R using 'igraph' or using a free network analysis software called 'Cytoscape', available for download from at \url{https://cytoscape.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Uiowa-Applied-Topology/mappeR}
  \item Report bugs at \url{https://github.com/Uiowa-Applied-Topology/mappeR/issues}
}

}
\author{
\strong{Maintainer}: George Clare Kennedy \email{george-clarekennedy@uiowa.edu}

}
\keyword{internal}
