% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappeR-package.R
\docType{package}
\name{mappeR-package}
\alias{mappeR}
\alias{mappeR-package}
\title{mappeR: Construct and Visualize TDA Mapper Graphs}
\description{
An implementation of various flavors of the mapper algorithm (from the field of topological data analysis) in R, originally developed by Singh, Mémoli and Carlsson (2007) (link to paper [here](https://research.math.osu.edu/tgda/mapperPBG.pdf), along with other tools for exploring the output graph in more detail. This can be done either by exporting the raw statistics, using the \pkg{igraph} package, or letting the package pass the data to Cytoscape (sold separately for free [here](https://cytoscape.org/)) for further analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Uiowa-Applied-Topology/mappeR}
  \item Report bugs at \url{https://github.com/Uiowa-Applied-Topology/mappeR/issues}
}

}
\author{
\strong{Maintainer}: George Clare Kennedy \email{george-clarekennedy@uiowa.edu}

}
\keyword{internal}
