% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mapper.R
\name{mapper}
\alias{mapper}
\title{Apply a function to each element of a vector, and its index via futures}
\usage{
mapper(workers = parallel::detectCores() - 2, ...)
}
\arguments{
\item{workers}{default value is the \verb{total cores - 2}}

\item{...}{Additional arguments passed on to the mapped function.

We now generally recommend against using \code{...} to pass additional
(constant) arguments to \code{.f}. Instead use a shorthand anonymous function:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Instead of
x |> map(f, 1, 2, collapse = ",")
# do:
x |> map(\\(x) f(x, 1, 2, collapse = ","))
}\if{html}{\out{</div>}}

This makes it easier to understand which arguments belong to which
function and will tend to yield better error messages.}
}
\value{
parameters
}
\description{
These functions work exactly the same as \code{\link[purrr:map]{purrr::map()}} functions,
but allow you to map in parallel.
}
