% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{mapper}
\alias{mapper}
\title{Run the mapper algorithm.}
\usage{
mapper(data, dists, filter, cover, clustering_method = NA)
}
\arguments{
\item{data}{A dataframe.}

\item{dists}{A matrix representing pairwise distances between datapoints in the dataframe.}

\item{filter}{A function \eqn{f:\ }\code{data} \eqn{\to\mathbb{R}^n}.}

\item{cover}{A list of subsets of \eqn{\mathbb{R}^n}.}

\item{clustering_method}{How to cluster datapoints in level sets of the filter function. Can be nothing, in which case the level set will count as one big cluster.}
}
\value{
something
}
\description{
Run the mapper algorithm.
}
