% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_studio.R
\name{mapper_object_to_igraph}
\alias{mapper_object_to_igraph}
\title{make igraph}
\usage{
mapper_object_to_igraph(mapperobject)
}
\arguments{
\item{mapperobject}{mapper object generated by mappeR}
}
\value{
an igraph object
}
\description{
make igraph
}
