% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\docType{class}
\name{MapperRef}
\alias{MapperRef}
\title{Mapper Reference Class (R6) implementation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MapperRef
}
\description{
Composes a set of classes to perform Mapper efficiently.
}
\author{
Matt Piekenbrock
}
\keyword{datasets}
