% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\encoding{UTF-8}
\name{MapperRef}
\alias{MapperRef}
\alias{Mapper}
\title{Mapper Reference Class (R6)}
\value{
\code{\link{MapperRef}} instance equipped with methods for building the mapper.
}
\description{
`MapperRef` is an \link{R6} class built for parameterizing and computing mappers efficiently.
}
\section{Details}{
 To create a new \code{MapperRef} instance object, instantiate the class with the \code{\link[R6:R6Class]{R6 new}} operator. 
Instantiation of a \code{MapperRef} objects requires a data matrix, or a function returning one. \cr
The primary output of the Mapper method is a simplicial complex. 
\cr 
The underlying complex does not need to be modified by the user, i.e. is completely maintained by \code{\link{MapperRef}} methods 
(e.g. \code{\link{construct_nerve}}, \code{\link{construct_k_skeleton}}, etc.).
}

\section{Usage}{

\preformatted{m = MapperRef$new(X)}
}

\section{Arguments}{

\itemize{
   \item{\strong{X}}: The data, either as a matrix, or a function that returns a matrix.
}
}

\section{Fields}{

\itemize{
  \item{\strong{X}}: The data, as a function. Evaluation returns the data as a matrix.
  \item{\strong{filter}}: The filter, as a function. Evaluation returns the filtered data as a matrix.  
  \item{\strong{cover}}: The cover (a \code{\link{CoverRef}} derived object). 
  \item{\strong{clustering_algorithm}}: The clustering algorithm to use in the pullback.
  \item{\strong{measure}}: Distance measure to use to compute distances in ambient space. See \code{use_distance_measure} for more details.     
  \item{\strong{pullback}}: Mapping between the sets in the cover (by index) and the vertices (by id).  
  \item{\strong{vertices}}: The mapper vertices. 
  \item{\strong{simplicial_complex}}: A \code{\link[simplextree:simplextree]{simplex tree}} object.
}
}

\section{Methods}{

\itemize{
  \item{\code{\link{use_filter}}: Specifies the filter.}
  \item{\code{\link{use_distance_measure}}: Specifies the distance measure.}
  \item{\code{\link{use_cover}}: Specifies the cover. Must be a \code{\link{CoverRef}} object.}
  \item{\code{\link{use_clustering_algorithm}}: Specifies the algorithm to decompose the pullback.}
  \item{\code{\link{construct_pullback}}: Decomposes the preimages into connected components.}
  \item{\code{\link{construct_nerve}}: Constructs simplices at a given dimension.}
  \item{\code{\link{construct_k_skeleton}}: Constructs the simplicial complex up to a given dimension.}
  \item{\code{\link{as_igraph}}: Converts the 1-skeleton to an \code{\link[igraph:igraph]{igraph object}}.}
  \item{\code{\link{as_pixiplex}}: Converts the simplicial complex to a \code{\link[pixiplex:pixiplex]{pixiplex object}}.}
  \item{\code{\link{exportMapper}}: Exports the core information of the mapper construction.}
}
}

\section{More information}{

Full documentation available \href{https://peekxc.github.io/Mapper}{online}.
}

\references{
Gurjeet Singh, Facundo Mémoli, and Gunnar Carlsson. "Topological methods for the analysis of high dimensional data sets and 3d object recognition." SPBG. 2007.
}
\author{
Matt Piekenbrock, \email{matt.piekenbrock@gmail.com}
}
