% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{mapperVis}
\alias{mapperVis}
\title{mapperVis - shiny server module}
\usage{
mapperVis(input, output, session, mapper, X, output_type)
}
\description{
Returns a set of reactiveValues for configuring the mapper visualization
}
