% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\docType{data}
\name{MapperVis}
\alias{MapperVis}
\title{MapperVis}
\format{An object of class \code{module} (inherits from \code{list}) of length 8.}
\usage{
MapperVis
}
\description{
'MapperVis' is a \code{\link[modules]{module}} containing \code{\link[shiny::callModule]{shiny modules}}
that simplify making shiny applications with Mapper.
}
\keyword{datasets}
