% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{multiscale}
\alias{multiscale}
\title{Computes all unique Mapper constructions
Only supports rectangular cover}
\usage{
multiscale(mapper_obj)
}
\description{
Computes all unique Mapper constructions
Only supports rectangular cover
}
