% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{multiscale}
\alias{multiscale}
\title{Multiscale Mapper}
\usage{
multiscale(
  m,
  max_dim = 1L,
  max_overlap = 50,
  time = c("integer", "measure", "overlap"),
  f = NULL,
  stats = FALSE
)
}
\arguments{
\item{m}{The initial mapper to start with.}

\item{max_dim}{The maximum dimension of the simplices to consider.}

\item{max_overlap}{The maximum overlap to consider. The multiscale construction will build only up to this value.}

\item{time}{The notion of 'time' to measure birth/death times with. See details.}

\item{f}{Function to evaluate at evry index. The mapper instance is passed as the first argument.}

\item{stats}{Collect statistics on number of operations required to compute the structure? Defaults to false.}
}
\description{
Computes the Multiscale Mapper construction. See details.
}
\details{
This method constructs a sequence of Mappers connected by simplicial maps, which can be readily 
passed to e.g. the \code{Simpers} library to produce persistence diagrams. 
Because constructing a sequence of Mappers is a computationally expensive operation, a special cover
indexing structure is used to express the computation in terms of elementary simplicial maps. This method requires 
a fixed interval cover with a constant number of open sets, and uses a specific clustering algorithm. \cr
\cr
The \code{time} parameter may be specified as "integer", "measure", or (average) "overlap". 
Each \code{time} unit is monotonically related to each other, however they may produce visually different persistence diagrams.
}
