% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartography_chamber.R
\name{next_triangular}
\alias{next_triangular}
\title{Triangular Number Triangulator}
\usage{
next_triangular(x)
}
\arguments{
\item{x}{A positive integer.}
}
\value{
The index of the next greatest or equal triangular number to \eqn{x}.
}
\description{
Given a number, find the cardinal order of the next closest triangular number.
}
