% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_pmapper.R
\name{pmapper_chr}
\alias{pmapper_chr}
\title{pmapper_chr}
\usage{
pmapper_chr(workers = (parallelly::availableCores() - 2), ...)
}
\arguments{
\item{workers}{multicore numbers: default value is the \verb{total cores - 2}}
}
\description{
pmapper_chr
}
