% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arboretum.R
\name{process_dendrograms}
\alias{process_dendrograms}
\title{Cut many dendrograms}
\usage{
process_dendrograms(dends)
}
\arguments{
\item{dends}{A list of dendrograms to be cut.}
}
\value{
A list of named vectors (one per dendrogram) whose names are data point names and whose values are cluster labels. This function determines a global minimum threshold based on the longest branches in all the input dendrograms, and uses that as a heuristic to gauge if the best number of clusters is 1, or the value obtained by cutting the longest branch.
}
\description{
Cut many dendrograms
}
