% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_factory.R
\name{process_dendrograms}
\alias{process_dendrograms}
\title{Cut many dendrograms}
\usage{
process_dendrograms(dends)
}
\arguments{
\item{dends}{A list of dendrograms to be cut.}
}
\value{
A list of named vectors (one per dendrogram) whose names are data point names and whose values are cluster labels.
}
\description{
Cut many dendrograms
}
\details{
This function uses a value of 10 percent of the tallest branch across dendrograms as a threshold for \link{cut_dendrogram}.
}
