% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterer_farm.R
\name{process_dendrograms}
\alias{process_dendrograms}
\title{Cut many dendrograms at specified cut heights.}
\usage{
process_dendrograms(dends, cut_heights)
}
\arguments{
\item{dends}{A list of dendrograms to be cut.}

\item{cut_heights}{A list of cut heights which corresponds to each dendrogram in \code{dends}.}
}
\value{
A list of named vectors (one per dendrogram) whose names are data point names and whose values are cluster labels.
}
\description{
Cut many dendrograms at specified cut heights.
}
