% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watering_hole.R
\name{run_cluster_machine}
\alias{run_cluster_machine}
\title{Ship data off to the clustering goblins}
\usage{
run_cluster_machine(dist_mats, method)
}
\arguments{
\item{dist_mats}{A list of distance matrices of each bin that is to be clustered.}

\item{method}{A string that suggests how the goblins will handle the data.}
}
\value{
A list containing named vectors (one per bin), whose names are data point names and whose values are cluster labels (within each bin)
}
\description{
This function tells the computer to look away for a second, so the goblins come and cluster your data while it's not watching.
}
