% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_methods.R
\name{run_link}
\alias{run_link}
\title{Perform agglomerative clustering on a single distance matrix.}
\usage{
run_link(dist, method)
}
\arguments{
\item{dist}{A distance matrix.}

\item{method}{A string to pass to \link{hclust} to determine clustering method.}
}
\value{
A dendrogram generated by \code{fastcluster}.
}
\description{
Perform agglomerative clustering on a single distance matrix.
}
