% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_factory.R
\name{run_link}
\alias{run_link}
\title{Perform single linkage clustering}
\usage{
run_link(dist, method)
}
\arguments{
\item{dist}{A distance matrix.}

\item{method}{A string to pass to \code{fastcluster} to determine clustering method.}
}
\value{
A dendrogram generated by \code{fastcluster}.
}
\description{
Perform single linkage clustering
}
