% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{simplex_tree}
\alias{simplex_tree}
\title{Simplex Tree}
\usage{
simplex_tree(n_vertices)
}
\value{
A queryable simplex tree.
}
\description{
Simplex tree class exposed as an Rcpp Module.
}
\details{
A simplex tree is a trie specialized for storing simplicial complexes.
}
\section{Fields}{

\describe{
\item{\code{n_vertices}}{Number of 0-simplexes to initialize the simplex tree with.}
}}

\references{
Mehta, Dinesh P., and Sartaj Sahni. "Chapter 18: Interval, Segment, Range, and Priority Search Trees." Handbook of data structures and applications. Chapman and Hall/CRC, 2004.
}
