% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{simplicial_complex}
\alias{simplicial_complex}
\title{Simplicial Complex}
\description{
The relational information of the Mapper construction.
}
\details{
The primary output of the Mapper method is a simplicial complex. With \code{\link{MapperRef}} objects, 
the simplicial complex is stored as a \code{\link[simplextree]{simplextree}}. 
\cr 
The underlying complex is completely maintained by \code{\link{MapperRef}} methods 
(e.g. \code{\link{compute_vertices}}, etc.). The complex may also be modified directly
via \code{\link[simplextree]{simplextree}} methods, however this is not recommended.
}
