% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{simplicial_complex}
\alias{simplicial_complex}
\title{Simplicial Complex}
\description{
The simplicial complex representing the relational information of the Mapper construction. 
With \code{\link{MapperRef}} objects, the simplicial complex is storied as a \code{\link[simplextree]{simplextree}}.
The underlying complex cannot be overwritten directly: use the given package functions (e.g. \code{\link{compute_vertices}}, etc.)
to make updates to the complex, or alternatively, modify the simplicial complex directly (not recommended).
}
