% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_factory.R
\name{subset_dists}
\alias{subset_dists}
\title{Subset a distance matrix}
\usage{
subset_dists(patch, dists)
}
\arguments{
\item{patch}{A list of names of data points.}

\item{dists}{A distance matrix for data points in the patch, possibly including extra points.}
}
\description{
Subset a distance matrix
}
