% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_factory.R
\name{subset_dists}
\alias{subset_dists}
\title{Distance Matrix Splicer}
\usage{
subset_dists(patch, dists)
}
\arguments{
\item{patch}{A list of names of data points.}

\item{dists}{A \code{dist} object for data points in the patch, possibly including extra points.}
}
\description{
Subset a \code{dist} object.
}
