% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{use_cover}
\alias{use_cover}
\title{Selects one of the available covering methods to use.}
\arguments{
\item{filter_values}{(n x d) numeric matrix of values giving the results of the map.}

\item{typename}{The name of the cover to use. Accepts any one of the typenames printed in the \code{available_covers}.}

\item{...}{Additional parameter values to pass to the covering generators initialize method.}
}
\description{
Convenience method to select, parameterize, and construct a cover and associate it with the calling objects \code{cover} field.
}
\details{
The cover is automatically constructed before being assigned to the calling Mappers \code{$cover} field.  
If this is undesirable, create the cover using the appropriate R6 class generators and assign to \code{$cover} field directly.
}
