% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{use_data}
\alias{use_data}
\title{Sets the data}
\arguments{
\item{data}{The data to cluster on with mapper. See details.}

\item{...}{additional parameters to pass to the filter function.}
}
\description{
Sets the data matrix to associate with the mapper.
}
\details{
Sets the data for the mapper construction. \code{data} can either be any input accepted
by the proxy package or a 'dist' object, 
or a function that returns a matrix of coordinate values. \cr
\cr
Alternatively, if \code{data} can also be a string giving the name of an existing data set listed in \code{data(package="Mapper")}.
}
