% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{use_distance_measure}
\alias{use_distance_measure}
\title{Sets the distance measure to use}
\arguments{
\item{measure}{The distance measure to use (string).}
}
\description{
Sets the distance measure to use
}
\details{
Specifies the distance measure to use to compute distances when decomposing the preimages. 
\code{measure} must be either one of ["euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"], or, 
if the \pkg{proxy} package is installed, any measure in \code{proxy::pr_DB$get_entry_names()}.
}
