% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperRef.R
\name{use_distance_measure}
\alias{use_distance_measure}
\title{Assigns a distance measure}
\arguments{
\item{measure}{The distance measure to use (string).}

\item{...}{Extra parameters passed to the distance function. See details.}
}
\description{
Assigns a distance measure to the \code{\link{MapperRef}} instance to use when compute distances in the data space.
}
\details{
The supplied \code{measure} must be either one of ["euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"], or, 
if the \pkg{proxy} package is installed, any measure in \code{proxy::pr_DB$get_entry_names()}.\cr
\cr
Additional parameters passed via \code{...} are passed to either \code{\link[stats]{dist}} or, if installed, 
\code{\link[parallelDist]{parallelDist}}. The former does not work with any measure in the \strong{proxy} package, 
whereas the latter should.
}
\section{Usage}{
 
\preformatted{$use_distance_measure(measure)}
}

\examples{
data(noisy_circle)
m <- MapperRef$new(noisy_circle)
m$use_filter(noisy_circle[,1])
m$use_cover("fixed interval", number_intervals = 5, percent_overlap = 25)

## Constructs clusters with euclidean metric (default)
m$use_distance_measure("euclidean")
m$construct_pullback() 

## Constructs clusters with p-norm (p = 1)
m$use_distance_measure("Minkowski", p = 1L)
m$construct_pullback() 

\dontrun{
## To see list of available measures, use:
proxy::pr_DB$get_entry_names()
}
}
\seealso{
\code{\link[parallelDist]{parDist}} \code{\link[proxy]{pr_DB}}
}
