% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\name{use_filter}
\alias{use_filter}
\title{Sets the filter}
\arguments{
\item{filter}{either the filter name to use, a matrix, or a function. See details.}

\item{...}{additional parameters to pass to the filter function.}
}
\description{
Sets the map, or \emph{filter}, to associate with the instance
}
\details{
\code{filter} must be either matrix of coordinate values, a function that returns a matrix of 
coordinate values, or a string selecting from one of the predefined filters below. \cr
\cr 
If a matrix is given (or function returning a matrix), the convention here is that each row 
represents a point and each column a dimension. \cr
\cr
Here's a list of the pre-defined filtering methods supported:
\itemize{
  \item{\strong{PC}}{ Principle components (with \code{\link[stats:prcomp]{prcomp}})}
  \item{\strong{IC}}{ Independent components (with \code{\link[fastICA:fastICA]{fastICA}})}
  \item{\strong{ECC}}{ Eccentricity (internal, change the norm by passing one of \eqn{p} = [1, 2, Inf])}
  \item{\strong{KDE}}{ Kernel Density Estimate (with \code{\link[ks:ks]{kde}})}
  \item{\strong{DTM}}{ Distance to measure (with \code{\link[TDA:dtm]{dtm}})}
  \item{\strong{MDS}}{ Classic (Metric) Multidimensional Scaling (with \code{\link[stats:cmdscale]{cmdscale}})}
  \item{\strong{ISOMAP}}{ Isometric feature mapping (with \code{\link[vegan:isomap]{isomap}})}
  \item{\strong{LE}}{ Laplacian Eigenmaps (with \code{\link[geigen:geigen]{geigen}})}
  \item{\strong{UMAP}}{ Uniform Manifold Approximation and Projection (with \code{\link[umap:umap]{umap}})}
}
Nearly all the pre-configured filters essentially call functions in other packages with
somewhat reasonable default parameters to perform the mapping. Any parameters supplied to \code{...} 
are passed to the corresponding package function linked above, which override any default parameters. 
If the package needed to compute the filter is not installed, a prompt is given asking the user 
whether they would like to install it. \cr
\cr
\strong{NOTE:} The predefined filters are meant to be used for exploratory or illustrative purposes only---this function 
is \emph{not} meant to act as a comprensive interface to the functions each filter corresponds too.
}
